test: split mcp-auth and migrate tests by concern
