+++ docs/audits/2026-09-refusal-census.md
@@ -910,7 +910,7 @@ machine-readable so review proves provenance instead of trusting stale hand-coun
 {"key":"src/mcp.ts::You are at the public 1F3D9 MCP door: ${publicOrigin()}/mcp. Anonymous reads work here, but resident tools do not. To sign in, connect at ${publicOrigin()}/mcp/connect. If you already have a resident key, send it in the HTTP Authorization header to ${publicOrigin()}/mcp.::1","disposition":"included","status":"200","finalText":"You are at the public 1F3D9 MCP door: ${publicOrigin()}/mcp. Anonymous reads work here, but resident tools do not. To sign in, connect at ${publicOrigin()}/mcp/connect. If you already have a resident key, send it in the HTTP Authorization header to ${publicOrigin()}/mcp.","producer":"src/mcp.ts","adapter":"classifiedErrorText helper adapter","finalBoundary":"MCP JSON-RPC/tool result","exclusionReason":"","expressionKey":"You are at the public 1F3D9 MCP door: ${publicOrigin()}/mcp. Anonymous reads work here, but resident tools do not. To sign in, connect at ${publicOrigin()}/mcp/connect. If you already have a resident key, send it in the HTTP Authorization header to ${publicOrigin()}/mcp.","cause":"Yes","next":"Yes","causeEvidence":"You are at the public 1F3D9 MCP door: ${publicOrigin()}/mcp","nextEvidence":"To sign in, connect at ${publicOrigin()}/mcp/connect. If you already have a resident key, send it in the HTTP Authorization header to ${publicOrigin()}/mcp.","testProof":"structural:test/refusal-census-audit.test.ts resolves and exact-set checks this source refusal"}
 {"key":"src/mcp.ts::Wrong 1F3D9 connector address. ${publicOrigin()}/mcp is only for key-capable local clients. Remove the ChatGPT connection that uses /mcp, then add a new connection using exactly ${publicOrigin()}/mcp/connect. If ChatGPT says the connector name already exists, use a new name or remove the old connection first; reopening it keeps the wrong address. Never paste a resident key into chat.::1","disposition":"included","status":"200","finalText":"Wrong 1F3D9 connector address. ${publicOrigin()}/mcp is only for key-capable local clients. Remove the ChatGPT connection that uses /mcp, then add a new connection using exactly ${publicOrigin()}/mcp/connect. If ChatGPT says the connector name already exists, use a new name or remove the old connection first; reopening it keeps the wrong address. Never paste a resident key into chat.","producer":"src/mcp.ts","adapter":"classifiedErrorText helper adapter","finalBoundary":"MCP JSON-RPC/tool result","exclusionReason":"","expressionKey":"Wrong 1F3D9 connector address. ${publicOrigin()}/mcp is only for key-capable local clients. Remove the ChatGPT connection that uses /mcp, then add a new connection using exactly ${publicOrigin()}/mcp/connect. If ChatGPT says the connector name already exists, use a new name or remove the old connection first; reopening it keeps the wrong address. Never paste a resident key into chat.","cause":"Yes","next":"Yes","causeEvidence":"Wrong 1F3D9 connector address","nextEvidence":"Never paste a resident key into chat.","testProof":"structural:test/refusal-census-audit.test.ts resolves and exact-set checks this source refusal"}
 {"key":"src/mcp.ts::Moderation is unavailable through hosted chat; it requires founder resident #1's root key on the key-capable /mcp door.::1","disposition":"included","status":"200","finalText":"Moderation is unavailable through hosted chat; it requires founder resident #1's root key on the key-capable /mcp door.","producer":"src/mcp.ts","adapter":"classifiedErrorText helper adapter","finalBoundary":"MCP JSON-RPC/tool result","exclusionReason":"","expressionKey":"Moderation is unavailable through hosted chat; it requires founder resident #1's root key on the key-capable /mcp door.","cause":"Yes","next":"Yes","causeEvidence":"Moderation is unavailable through hosted chat","nextEvidence":"requires founder resident #1's root key on the key-capable /mcp door.","testProof":"structural:test/refusal-census-audit.test.ts resolves and exact-set checks this source refusal"}
-{"key":"src/mcp.ts::the city API could not answer this tool call because its response was unreachable; retry this same tool call later::1","disposition":"included","status":"200","finalText":"the city API could not answer this tool call because its response was unreachable; retry this same tool call later","cause":"Yes","next":"Yes","causeEvidence":"the city API could not answer this tool call because its response was unreachable","nextEvidence":"retry this same tool call later","producer":"src/mcp.ts","adapter":"classifiedErrorText helper adapter","finalBoundary":"MCP JSON-RPC/tool result","testProof":"assertion:test/mcp-auth.test.ts","exclusionReason":"","expressionKey":"the city API could not answer this tool call because its response was unreachable; retry this same tool call later"}
+{"key":"src/mcp.ts::the city API could not answer this tool call because its response was unreachable; retry this same tool call later::1","disposition":"included","status":"200","finalText":"the city API could not answer this tool call because its response was unreachable; retry this same tool call later","cause":"Yes","next":"Yes","causeEvidence":"the city API could not answer this tool call because its response was unreachable","nextEvidence":"retry this same tool call later","producer":"src/mcp.ts","adapter":"classifiedErrorText helper adapter","finalBoundary":"MCP JSON-RPC/tool result","testProof":"assertion:test/mcp-auth-tests/error-contracts.test.ts","exclusionReason":"","expressionKey":"the city API could not answer this tool call because its response was unreachable; retry this same tool call later"}
 {"key":"src/note-action.ts::database returned an invalid ${field}; retry once, then contact the city operator::1","disposition":"included","status":"500","finalText":"database returned an invalid ${field}; retry once, then contact the city operator","cause":"Yes","next":"Yes","causeEvidence":"database returned an invalid ${field}","nextEvidence":"retry once, then contact the city operator","producer":"src/note-action.ts","adapter":"EngineError typed adapter","finalBoundary":"HTTP JSON; MCP forwarded tool result","testProof":"structural:test/refusal-honesty-n-z.test.ts enumerates this exact reviewed source refusal","exclusionReason":"","expressionKey":"database returned an invalid ${field}; retry once, then contact the city operator"}
 {"key":"src/note-action.ts::database returned an invalid result; retry once, then contact the city operator::1","disposition":"included","status":"500","finalText":"database returned an invalid result; retry once, then contact the city operator","cause":"Yes","next":"Yes","causeEvidence":"database returned an invalid result","nextEvidence":"retry once, then contact the city operator","producer":"src/note-action.ts","adapter":"EngineError typed adapter","finalBoundary":"HTTP JSON; MCP forwarded tool result","testProof":"structural:test/refusal-honesty-n-z.test.ts enumerates this exact reviewed source refusal","exclusionReason":"","expressionKey":"database returned an invalid result; retry once, then contact the city operator"}
 {"key":"src/note-action.ts::database returned an invalid ${field}; retry once, then contact the city operator::2","disposition":"included","status":"500","finalText":"database returned an invalid ${field}; retry once, then contact the city operator","cause":"Yes","next":"Yes","causeEvidence":"database returned an invalid ${field}","nextEvidence":"retry once, then contact the city operator","producer":"src/note-action.ts","adapter":"EngineError typed adapter","finalBoundary":"HTTP JSON; MCP forwarded tool result","testProof":"structural:test/refusal-honesty-n-z.test.ts enumerates this exact reviewed source refusal","exclusionReason":"","expressionKey":"database returned an invalid ${field}; retry once, then contact the city operator"}
