+++ docs/audits/openai-codex_identity_Audit_Findings.md
@@ -214,7 +214,7 @@ The disposable proofs use synthetic values and replace `globalThis.fetch` before
 - **Evidence level:** E2
 - **Status:** Likely
 - **Rule or parameter:** The code states that publishing a resident credential transfers authority; every unauthenticated output boundary must therefore prevent stored credentials from being returned, including data created before the write guard.
-- **Location:** `src/input.ts:11`; `src/world.ts:97`; `src/world.ts:167`; `src/society.ts:163`; `src/society.ts:460`; `src/moderation-store.ts:114`; `src/mcp.ts:647`; `src/mcp.ts:674`; `test/mcp-auth.test.ts:403`; commit `dbcc133292ac652f940086d86b8d755d76bd7ddf`
+- **Location:** `src/input.ts:11`; `src/world.ts:97`; `src/world.ts:167`; `src/society.ts:163`; `src/society.ts:460`; `src/moderation-store.ts:114`; `src/mcp.ts:647`; `src/mcp.ts:674`; `test/mcp-auth-tests/redaction-and-passivity.test.ts:20`; commit `dbcc133292ac652f940086d86b8d755d76bd7ddf`
 - **User or business harm:** An anonymous reader can receive a still-valid root key or connector credential from a historical/imported note, thing, place description, or agreement body and immediately gain resident authority.
 - **Evidence:** Raw REST routes return stored authored fields after moderation-only processing. Hosted MCP alone scans responses and redacts historical credential-bearing note bodies or withholds the response. A synthetic fake-Neon `/api/note/:id` proof returned a synthetic resident key unchanged. The fresh skeptic kept E2 because production rows were not inspected and new writes are now rejected.
 - **Safe reproduction:** EVD-007 injects one synthetic historical note through a fake database response, calls the public route without authentication, and checks only whether the synthetic string is present; no live row or real credential is used.
