+++ docs/audits/openai-codex_world_Audit_Findings.md
@@ -148,9 +148,9 @@ Long command output was summarized rather than pasted. No command read a real be
 - **Evidence level:** E3
 - **Status:** Confirmed
 - **Rule or parameter:** Resident root keys and OAuth credentials must never be published or returned through public output; current input code calls publication an unwitting transfer of authority (`src/input.ts:11-23`).
-- **Location:** `src/mcp.ts:646`, `src/mcp.ts:855`, `src/society.ts:163`, `src/world.ts:141`, `src/window.ts:503`, `src/window.ts:740`, `src/moderation-store.ts:97`, `test/mcp-auth.test.ts:403`
+- **Location:** `src/mcp.ts:646`, `src/mcp.ts:855`, `src/society.ts:163`, `src/world.ts:141`, `src/window.ts:503`, `src/window.ts:740`, `src/moderation-store.ts:97`, `test/mcp-auth-tests/redaction-and-passivity.test.ts:20`
 - **User or business harm:** If even one pre-guard note contains an unrotated resident root key, any anonymous reader can impersonate that resident and act on its land, things, agreements, and transfers. Exposed access/refresh credentials can also grant or extend hosted access. Actual production presence is unknown, but the read-layer behavior is confirmed and the source explicitly anticipates historical rows.
-- **Evidence:** New writes reject credential-shaped text at `society.ts:184`, but `mcp.ts:646-650` says historical notes may predate that guard. Its redactor runs only when `hostedChat` is true at `mcp.ts:855-862`. Public note, place-note, and window queries select stored `note.body` and pass it through moderation; `moderatePublicRows` changes content only when an explicit latest moderation action is `remove`. The repository fixture at `test/mcp-auth.test.ts:403-501` verifies hosted redaction for pre-guard root/access-token-shaped notes. A synthetic legacy-MCP reproduction returned the fake credential-shaped body unchanged.
+- **Evidence:** New writes reject credential-shaped text at `society.ts:184`, but `mcp.ts:646-650` says historical notes may predate that guard. Its redactor runs only when `hostedChat` is true at `mcp.ts:855-862`. Public note, place-note, and window queries select stored `note.body` and pass it through moderation; `moderatePublicRows` changes content only when an explicit latest moderation action is `remove`. The repository fixture at `test/mcp-auth-tests/redaction-and-passivity.test.ts:20-122` verifies hosted redaction for pre-guard root/access-token-shaped notes. A synthetic legacy-MCP reproduction returned the fake credential-shaped body unchanged.
 - **Safe reproduction:** A local Hono app mounted `mcp()` in legacy mode over a synthetic public place response containing `1f3d9_sk_` plus generated fake hex. Anonymous `look` returned status 200 and success; `synthetic_credential_returned` was `true` and `redaction_marker_present` was `false`. No production note or real credential was read, stored, or printed.
 - **Connection traced:** Historical `notes.body` → `GET /api/note/:id`, embedded `GET /api/place/:id`, or window collection query → moderation overlay only → anonymous JSON/window output; legacy MCP `look` simply forwards that raw route response because the hosted-only safeguard is skipped.
 - **Root cause:** Legacy-secret handling was added as a hosted-MCP response wrapper instead of at the shared public-note shaping boundary. Current write rejection prevents new rows but does not protect older stored rows or every read surface.
