+++ test/mcp-auth.test.ts
@@ -1,2174 +1,18 @@
-// Release 1 OAuth contract tests use an in-memory Hono app only.
-// No live database, deployment, wallet, or network service is touched.
-import test from 'node:test'
-import assert from 'node:assert/strict'
-import { Hono } from 'hono'
-import {
-  auth,
-  authPassive,
-  setOAuthResidentResolver,
-  setPassiveOAuthResidentResolver,
-} from '../src/core.ts'
-import {
-  LATER_HOLDER_CURSOR_LENGTH,
-  LATER_HOLDER_CURSOR_PATTERN,
-} from '../src/later-holder.ts'
-import { mcp } from '../src/mcp.ts'
-import { PUBLIC_EVENT_KINDS } from '../src/public-events.ts'
-
-const PUBLIC_ORIGIN = 'https://1f3d9.com'
-const LEGACY_SECRET = `1f3d9_sk_${'ab'.repeat(24)}`
-const OAUTH_ACCESS_TOKEN = `1f3d9_at_${'cd'.repeat(32)}`
-const RESOURCE_METADATA = `${PUBLIC_ORIGIN}/.well-known/oauth-protected-resource/mcp/connect`
-const FRONT_DOOR_POINTER =
-  'Lost? Read the city front door with the front_door tool, or at https://1f3d9.com/ if your client can open URLs.'
-const OAUTH_SCHEME = { type: 'oauth2', scopes: ['city:resident'] } as const
-const NOAUTH_SCHEME = { type: 'noauth' } as const
-
-process.env.PUBLIC_ORIGIN = PUBLIC_ORIGIN
-
-interface ToolDefinition {
-  name: string
-  title?: string
-  description: string
-  inputSchema: {
-    additionalProperties?: boolean
-    properties?: Record<string, unknown>
-    required?: string[]
-  }
-  annotations?: {
-    readOnlyHint?: boolean
-    destructiveHint?: boolean
-    idempotentHint?: boolean
-    openWorldHint?: boolean
-  }
-  securitySchemes?: unknown[]
-  _meta?: { securitySchemes?: unknown[] }
-}
-
-interface ToolResult {
-  isError: boolean
-  content: { type: string; text: string }[]
-  _meta?: { 'mcp/www_authenticate'?: string[] }
-}
-
-const EXISTING_TOOL_NAMES = [
-  'front_door', 'help', 'official_facts', 'physics', 'search', 'changes', 'look',
-  'browse', 'drawing', 'drawing_history', 'credit_preflight', 'buy_credit', 'found', 'place_edit',
-  'coin_trait', 'invent_kind', 'revise_kind', 'make', 'thing_edit', 'thing_upgrade',
-  'draw_self', 'act', 'laws', 'home', 'withdraw',
-  'list_world', 'claim_world', 'cancel_world', 'reconcile_world', 'credit_gift',
-  'payment_attempt', 'transfer',
-  'agree', 'open_agreement_accession', 'sign', 'say', 'flag', 'later_holder_items',
-  'mark_for_later', 'me', 'moderate',
-] as const
-const PUBLIC_ANONYMOUS_TOOL_NAMES = [
-  'front_door', 'help', 'official_facts', 'physics', 'search', 'changes', 'look', 'browse',
-  'drawing', 'drawing_history',
-] as const
-
-const TOOL_TITLES: Readonly<Record<(typeof EXISTING_TOOL_NAMES)[number], string>> = Object.freeze({
-  front_door: 'Read front door',
-  help: 'Read city help',
-  official_facts: 'Read official facts',
-  physics: 'Read city physics',
-  search: 'Search public records',
-  changes: 'Check public changes',
-  look: 'Look around',
-  browse: 'Browse public catalogs',
-  drawing: 'Read a drawing',
-  drawing_history: 'Read drawing history',
-  credit_preflight: 'Check one fee before confirming',
-  buy_credit: 'Buy city credit',
-  found: 'Found a place',
-  place_edit: 'Edit a place',
-  coin_trait: 'Coin a trait',
-  invent_kind: 'Invent a kind',
-  revise_kind: 'Revise a kind',
-  make: 'Make a thing',
-  thing_edit: 'Edit a thing',
-  thing_upgrade: 'Upgrade a thing',
-  draw_self: 'Draw myself',
-  act: 'Act in the city',
-    laws: 'Set regional laws',
-  home: 'Set home',
-  withdraw: 'Withdraw a thing',
-  list_world: 'List a world thing',
-  claim_world: 'Claim a world thing',
-  cancel_world: 'Cancel a world listing',
-  reconcile_world: 'Reconcile a world payment',
-  credit_gift: 'Accept or refuse a credit gift',
-  payment_attempt: 'Check a payment attempt',
-  transfer: 'Transfer property',
-  agree: 'Write an agreement',
-  open_agreement_accession: 'Open agreement accession',
-  sign: 'Sign an agreement',
-  say: 'Speak here',
-  flag: 'Flag illegal content',
-  later_holder_items: 'Check marked items',
-  mark_for_later: 'Mark or unmark a thing',
-  me: 'Check my status',
-  moderate: 'Moderate illegal content',
-})
-
-const PROTECTED_TOOL_NAMES = [
-  'credit_preflight', 'buy_credit', 'found', 'place_edit', 'coin_trait',
-  'invent_kind', 'revise_kind', 'make', 'thing_edit', 'thing_upgrade',
-  'draw_self', 'act', 'laws', 'home', 'withdraw', 'list_world',
-  'claim_world', 'cancel_world', 'reconcile_world', 'credit_gift', 'payment_attempt',
-  'transfer', 'agree',
-  'open_agreement_accession', 'sign', 'say', 'flag', 'later_holder_items',
-  'mark_for_later', 'me',
-] as const
-const HOSTED_TOOL_NAMES = [...PUBLIC_ANONYMOUS_TOOL_NAMES, ...PROTECTED_TOOL_NAMES] as const
-
-function setHostedChatFlag(enabled: boolean) {
-  process.env.HOSTED_CHAT_SIGNIN_ENABLED = enabled ? 'true' : 'false'
-}
-
-function createHarness() {
-  let forwardedAuthorization: string | undefined
-  let forwardedMethod: string | undefined
-  let forwardedBody: unknown
-  let noteCalls = 0
-
-  const city = new Hono()
-  city.get('/', c => c.text('connector-native front door\n'))
-  city.get('/api/help', c => c.json({ doors: ['City map: `look` starts here.'] }))
-  city.get('/api/official', c => c.json({ domain: PUBLIC_ORIGIN, token: null }))
-  city.get('/api/physics', c => c.json({ basic_actions: ['move'], max_effect_depth: 12 }))
-  city.get('/api/me', async c => {
-    forwardedAuthorization = c.req.header('authorization')
-    forwardedMethod = c.req.method
-    if (forwardedAuthorization === `Bearer ${LEGACY_SECRET}`) {
-      return c.json({
-        resident: { id: 49, handle: 'chatty' },
-        front_door: `${PUBLIC_ORIGIN}/`,
-        front_door_tool: 'front_door',
-      })
-    }
-    const resident = await auth(c)
-    if (resident) {
-      return c.json({ resident, front_door: `${PUBLIC_ORIGIN}/`, front_door_tool: 'front_door' })
-    }
-
-    c.header(
-      'WWW-Authenticate',
-      `Bearer resource_metadata="${RESOURCE_METADATA}", error="invalid_token", ` +
-        'error_description="Sign in to 1F3D9 to use resident tools."',
-    )
-    return c.json({ error: 'A valid resident sign-in is required.' }, 401)
-  })
-  city.post('/api/me', async c => {
-    forwardedAuthorization = c.req.header('authorization')
-    forwardedMethod = c.req.method
-    forwardedBody = await c.req.json()
-    if (forwardedAuthorization === `Bearer ${LEGACY_SECRET}` || await authPassive(c)) {
-      c.header('Cache-Control', 'no-store')
-      return c.json({ count: 1, question: 'approved question' })
-    }
-    return c.json({ error: 'A valid resident sign-in is required.' }, 401)
-  })
-  city.patch('/api/me/drawing', async c => {
-    forwardedAuthorization = c.req.header('authorization')
-    forwardedMethod = c.req.method
-    forwardedBody = await c.req.json()
-    if (forwardedAuthorization === `Bearer ${LEGACY_SECRET}` || await authPassive(c)) {
-      c.header('Cache-Control', 'no-store')
-      return c.json({ resident: { id: 49, handle: 'chatty', drawing: forwardedBody }, changed: true })
-    }
-    return c.json({ error: 'A valid resident sign-in is required.' }, 401)
-  })
-  city.get('/api/drawing/:type/:id', c => {
-    forwardedAuthorization = c.req.header('authorization')
-    forwardedMethod = c.req.method
-    return c.json({
-      type: c.req.param('type'),
-      id: Number(c.req.param('id')),
-      state: 'complete',
-      presentation_state: 'complete',
-      description: 'A public lantern.',
-      drawing: {
-        palette: ['#ad3f25'],
-        indices: Array.from({ length: 64 }, (_, index) => index === 0 ? 0 : null),
-      },
-      rows: ['0 . . . . . . .', ...Array.from({ length: 7 }, () => '. . . . . . . .')],
-      source: c.req.param('type') === 'kind' ? 'kind_base' : c.req.param('type'),
-    })
-  })
-  city.get('/api/drawing/:type/:id/history', c => {
-    forwardedAuthorization = c.req.header('authorization')
-    forwardedMethod = c.req.method
-    forwardedBody = { ...c.req.query() }
-    return c.json({
-      type: c.req.param('type'),
-      id: Number(c.req.param('id')),
-      revisions: [],
-      page: { limit: Number(c.req.query('limit') ?? 20), has_more: false, next_before: null },
-    })
-  })
-  city.post('/api/thing/:id/mark', async c => {
-    forwardedAuthorization = c.req.header('authorization')
-    forwardedMethod = c.req.method
-    forwardedBody = await c.req.json()
-    if (forwardedAuthorization === `Bearer ${LEGACY_SECRET}` || await authPassive(c)) {
-      c.header('Cache-Control', 'no-store')
-      return c.json({ thing_id: Number(c.req.param('id')), marked: true, changed: true })
-    }
-    return c.json({ error: 'A valid resident sign-in is required.' }, 401)
-  })
-  city.post('/api/note', c => {
-    noteCalls += 1
-    return c.json({ note: { id: 1 } }, 201)
-  })
-
-  const gateway = new Hono()
-  gateway.post('/mcp', c => mcp(c, city))
-  gateway.post('/mcp/connect', c => mcp(c, city, {
-    hostedChat: true,
-    forwardUnauthorizedStatus: false,
-  }))
-
-  return {
-    city,
-    gateway,
-    forwardedAuthorization: () => forwardedAuthorization,
-    forwardedMethod: () => forwardedMethod,
-    forwardedBody: () => forwardedBody,
-    noteCalls: () => noteCalls,
-  }
-}
-
-function createAuthenticatedLookHarness(payload: Record<string, unknown>) {
-  const city = new Hono()
-  city.get('/api/place/:id', c => {
-    if (![OAUTH_ACCESS_TOKEN, LEGACY_SECRET]
-      .some(secret => c.req.header('authorization') === `Bearer ${secret}`)) {
-      return c.json({ error: 'A valid resident sign-in is required.' }, 401)
-    }
-    return c.json(payload)
-  })
-  city.get('/api/thing/:id', c => {
-    if (![OAUTH_ACCESS_TOKEN, LEGACY_SECRET]
-      .some(secret => c.req.header('authorization') === `Bearer ${secret}`)) {
-      return c.json({ error: 'A valid resident sign-in is required.' }, 401)
-    }
-    return c.json(payload)
-  })
-  city.get('/api/note/:id', c => {
-    if (![OAUTH_ACCESS_TOKEN, LEGACY_SECRET]
-      .some(secret => c.req.header('authorization') === `Bearer ${secret}`)) {
-      return c.json({ error: 'A valid resident sign-in is required.' }, 401)
-    }
-    return c.json(payload)
-  })
-
-  const gateway = new Hono()
-  gateway.post('/mcp', c => mcp(c, city))
-  gateway.post('/mcp/connect', c => mcp(c, city, {
-    hostedChat: true,
-    forwardUnauthorizedStatus: false,
-  }))
-  return gateway
-}
-
-function createVisitOpeningHarness() {
-  const backingCalls: Array<{
-    method: string
-    path: string
-    authorization: string | null
-    body?: unknown
-  }> = []
-  const city = new Hono()
-  const record = (method: string, path: string, authorization: string | null, body?: unknown) => {
-    backingCalls.push({ method, path, authorization, ...(body === undefined ? {} : { body }) })
-  }
-
-  city.get('/', c => {
-    record(c.req.method, c.req.path, c.req.header('authorization') ?? null)
-    return c.text('connector-native front door\n')
-  })
-  city.get('/api/official', c => {
-    record(c.req.method, c.req.path, c.req.header('authorization') ?? null)
-    return c.json({ domain: PUBLIC_ORIGIN, token: null })
-  })
-  city.get('/api/me', async c => {
-    record(c.req.method, c.req.path, c.req.header('authorization') ?? null)
-    const resident = await auth(c)
-    if (!resident) return c.json({ error: 'A valid resident sign-in is required.' }, 401)
-    return c.json({
-      handle: resident.handle,
-      front_door: `${PUBLIC_ORIGIN}/`,
-      front_door_tool: 'front_door',
-    })
-  })
-  city.post('/api/action', async c => {
-    const body = await c.req.json()
-    record(c.req.method, c.req.path, c.req.header('authorization') ?? null, body)
-    const resident = await auth(c)
-    if (!resident) return c.json({ error: 'A valid resident sign-in is required.' }, 401)
-    return c.json({ action: { action: 'go_home', status: 'applied', actor: resident.handle } })
-  })
-
-  const gateway = new Hono()
-  gateway.post('/mcp/connect', c => mcp(c, city, {
-    hostedChat: true,
-    forwardUnauthorizedStatus: false,
-  }))
-  return { gateway, backingCalls }
-}
-
-async function rpc(
-  app: Hono,
-  method: string,
-  params?: Record<string, unknown>,
-  authorization?: string,
-  path = '/mcp/connect',
-) {
-  const headers: Record<string, string> = { 'Content-Type': 'application/json' }
-  if (authorization) headers.Authorization = authorization
-  const response = await app.request(path, {
-    method: 'POST',
-    headers,
-    body: JSON.stringify({ jsonrpc: '2.0', id: 1, method, params }),
-  })
-  assert.equal(response.status, 200)
-  return response.json() as Promise<Record<string, unknown>>
-}
-
-async function listTools(app: Hono, path = '/mcp/connect', authorization?: string): Promise<ToolDefinition[]> {
-  const response = await rpc(app, 'tools/list', undefined, authorization, path) as {
-    result: { tools: ToolDefinition[] }
-  }
-  return response.result.tools
-}
-
-function toolByName(tools: ToolDefinition[], name: string) {
-  const tool = tools.find(candidate => candidate.name === name)
-  assert.ok(tool, `tools/list should include ${name}`)
-  return tool
-}
-
-async function callTool(
-  app: Hono,
-  name: string,
-  arguments_: Record<string, unknown>,
-  authorization?: string,
-  path = '/mcp/connect',
-): Promise<ToolResult> {
-  const response = await rpc(app, 'tools/call', {
-    name,
-    arguments: arguments_,
-  }, authorization, path) as {
-    result?: ToolResult
-    error?: { message?: string }
-  }
-  assert.ok(response.result, response.error?.message ?? `${name} returned no tool result`)
-  return response.result
-}
-
-test('initialize defaults to the current MCP version and echoes an explicit current version', async () => {
-  setHostedChatFlag(true)
-  const { gateway } = createHarness()
-  const fallback = await rpc(gateway, 'initialize', {}) as {
-    result: { protocolVersion: string }
-  }
-  assert.equal(fallback.result.protocolVersion, '2025-11-25')
-
-  const current = await rpc(gateway, 'initialize', {
-    protocolVersion: '2025-11-25',
-  }) as { result: { protocolVersion: string } }
-  assert.equal(current.result.protocolVersion, '2025-11-25')
-})
-
-test('feature on advertises OAuth for resident tools and mixed auth for every public read', async () => {
-  setHostedChatFlag(true)
-  const { gateway } = createHarness()
-  const tools = await listTools(gateway)
-
-  for (const name of PROTECTED_TOOL_NAMES) {
-    assert.deepEqual(toolByName(tools, name).securitySchemes, [OAUTH_SCHEME], name)
-  }
-  for (const name of PUBLIC_ANONYMOUS_TOOL_NAMES) {
-    const tool = toolByName(tools, name)
-    assert.deepEqual(tool.securitySchemes, [NOAUTH_SCHEME, OAUTH_SCHEME], name)
-    assert.deepEqual(tool._meta?.securitySchemes, tool.securitySchemes, `${name}: compatibility mirror`)
-  }
-  assert.equal(tools.some(tool => tool.name === 'register'), false)
-  assert.equal(tools.some(tool => tool.name === 'moderate'), false)
-
-  const advertised = JSON.stringify(tools)
-  assert.doesNotMatch(advertised, /1f3d9_(?:at|rt|ac)_/i)
-  for (const forbiddenField of ['access_token', 'refresh_token', 'client_secret', 'code', 'session']) {
-    assert.equal(
-      tools.some(tool => forbiddenField in (tool.inputSchema.properties ?? {})),
-      false,
-      `${forbiddenField} must never be a tool argument`,
-    )
-  }
-})
-
-test('connector-native reference tools accept no arguments and are safe anonymous reads', async () => {
-  setHostedChatFlag(true)
-  const { gateway } = createHarness()
-  const tools = await listTools(gateway)
-  const readAnnotations = {
-    readOnlyHint: true,
-    destructiveHint: false,
-    idempotentHint: true,
-    openWorldHint: true,
-  }
-
-  for (const name of ['front_door', 'help', 'official_facts', 'physics'] as const) {
-    const tool = toolByName(tools, name)
-    assert.equal(tool.inputSchema.additionalProperties, false, `${name}: closed input`)
-    assert.deepEqual(tool.inputSchema.properties ?? {}, {}, `${name}: no arguments`)
-    assert.deepEqual(tool.inputSchema.required ?? [], [], `${name}: no required arguments`)
-    assert.deepEqual(tool.annotations, readAnnotations, `${name}: read-only annotations`)
-  }
-})
-
-test('connector-native reference tools execute anonymously with identical content on both MCP doors', async () => {
-  setHostedChatFlag(true)
-  const { gateway } = createHarness()
-  const expected = {
-    front_door: 'connector-native front door\n',
-    help: JSON.stringify({ doors: ['City map: `look` starts here.'] }),
-    official_facts: JSON.stringify({ domain: PUBLIC_ORIGIN, token: null }),
-    physics: JSON.stringify({ basic_actions: ['move'], max_effect_depth: 12 }),
-  } as const
-
-  for (const name of ['front_door', 'help', 'official_facts', 'physics'] as const) {
-    const legacy = await callTool(gateway, name, {}, undefined, '/mcp')
-    const hosted = await callTool(gateway, name, {}, undefined, '/mcp/connect')
-    assert.equal(legacy.isError, false, `/mcp: ${name}`)
-    assert.equal(hosted.isError, false, `/mcp/connect: ${name}`)
-    assert.equal(legacy.content[0]?.text, expected[name], `/mcp: ${name}`)
-    assert.equal(hosted.content[0]?.text, legacy.content[0]?.text, `${name}: identical doors`)
-  }
-})
-
-test('a hosted resident can open a visit through connector tools without a global web fetch', async () => {
-  setHostedChatFlag(true)
-  const resident = {
-    id: 49,
-    handle: 'chatty',
-    model: 'hosted-chat',
-    joined_at: '2026-08-13T00:00:00.000Z',
-    quota_day: '2026-08-13',
-    things_today: 0,
-    notes_today: 0,
-    agreement_actions_today: 0,
-  }
-  setOAuthResidentResolver(async token => token === OAUTH_ACCESS_TOKEN ? resident : null)
-  const originalFetch = globalThis.fetch
-  let globalFetchCalls = 0
-  globalThis.fetch = (async () => {
-    globalFetchCalls += 1
-    throw new Error('the hosted visit-opening sequence must not use global fetch')
-  }) as typeof fetch
-
-  try {
-    const { gateway, backingCalls } = createVisitOpeningHarness()
-    const authorization = `Bearer ${OAUTH_ACCESS_TOKEN}`
-    const frontDoor = await callTool(gateway, 'front_door', {}, authorization)
-    const officialFacts = await callTool(gateway, 'official_facts', {}, authorization)
-    const me = await callTool(gateway, 'me', {}, authorization)
-    const act = await callTool(gateway, 'act', {
-      action: 'move', to_place_id: 3, carry_thing_id: 41,
-    }, authorization)
-
-    assert.equal(frontDoor.isError, false)
-    assert.equal(frontDoor.content[0]?.text, 'connector-native front door\n')
-    assert.equal(officialFacts.isError, false)
-    assert.deepEqual(JSON.parse(officialFacts.content[0]?.text ?? '{}'), {
-      domain: PUBLIC_ORIGIN,
-      token: null,
-    })
-    assert.equal(JSON.parse(me.content[0]?.text ?? '{}').handle, 'chatty')
-    assert.equal(JSON.parse(act.content[0]?.text ?? '{}').action.status, 'applied')
-    assert.deepEqual(
-      backingCalls.map(call => [call.method, call.path]),
-      [
-        ['GET', '/'],
-        ['GET', '/api/official'],
-        ['GET', '/api/me'],
-        ['POST', '/api/action'],
-      ],
-    )
-    assert.deepEqual(
-      backingCalls.map(call => call.authorization),
-      [authorization, authorization, authorization, authorization],
-    )
-    assert.deepEqual(backingCalls.at(-1)?.body, {
-      action: 'move', to_place_id: 3, carry_thing_id: 41,
-    })
-    assert.equal(globalFetchCalls, 0)
-  } finally {
-    globalThis.fetch = originalFetch
-    setOAuthResidentResolver(null)
-  }
-})
-
-test('every advertised MCP tool has a short plain title on its exact door catalog', async () => {
-  assert.equal(EXISTING_TOOL_NAMES.length, 41)
-  assert.equal(HOSTED_TOOL_NAMES.length, 40)
-  for (const [hosted, path, authorization, expectedNames] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`, HOSTED_TOOL_NAMES],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`, EXISTING_TOOL_NAMES],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const { gateway } = createHarness()
-    const tools = await listTools(gateway, path, authorization)
-    assert.deepEqual(tools.map(tool => tool.name), expectedNames, `${path}: exact catalog`)
-    for (const tool of tools) {
-      assert.equal(tool.title, TOOL_TITLES[tool.name as keyof typeof TOOL_TITLES], `${path}: ${tool.name}`)
-      assert.match(tool.title ?? '', /^[A-Z][A-Za-z ]{2,39}$/u, `${path}: ${tool.name}`)
-    }
-  }
-})
-
-test('every authenticated MCP surface carries one quiet front-door pointer', async () => {
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const { gateway } = createHarness()
-    const initialized = await rpc(gateway, 'initialize', {}, authorization, path) as {
-      result: { instructions: string }
-    }
-    assert.equal(
-      initialized.result.instructions.split(FRONT_DOOR_POINTER).length - 1,
-      1,
-      `${path}: initialize`,
-    )
-
-    for (const tool of await listTools(gateway, path, authorization)) {
-      assert.equal(
-        tool.description.split(FRONT_DOOR_POINTER).length - 1,
-        1,
-        `${path}: ${tool.name}`,
-      )
-    }
-
-    const badCall = await rpc(
-      gateway,
-      'tools/call',
-      { name: 'me', arguments: { unsupported: true } },
-      authorization,
-      path,
-    ) as { result: { content: Array<{ text: string }> } }
-    const badCallBody = JSON.parse(badCall.result.content[0]!.text) as {
-      front_door?: string
-      front_door_tool?: string
-    }
-    assert.equal(badCallBody.front_door, 'https://1f3d9.com/')
-    assert.equal(badCallBody.front_door_tool, 'front_door')
-
-    const unknownMethod = await rpc(
-      gateway,
-      'city/unknown',
-      {},
-      authorization,
-      path,
-    ) as {
-      error: {
-        code: number
-        message: string
-        data?: { front_door?: string; front_door_tool?: string }
-      }
-    }
-    assert.equal(unknownMethod.error.code, -32601, `${path}: unknown method`)
-    assert.equal(
-      unknownMethod.error.message,
-      'method not found: city/unknown; call initialize, ping, tools/list, or tools/call',
-      `${path}: unknown method`,
-    )
-    assert.equal(
-      unknownMethod.error.data?.front_door,
-      'https://1f3d9.com/',
-      `${path}: unknown method front door`,
-    )
-    assert.equal(
-      unknownMethod.error.data?.front_door_tool,
-      'front_door',
-      `${path}: unknown method front-door tool`,
-    )
-
-    const unknownTool = await rpc(
-      gateway,
-      'tools/call',
-      { name: 'unknown_city_tool', arguments: {} },
-      authorization,
-      path,
-    ) as {
-      error: {
-        code: number
-        message: string
-        data?: { front_door?: string; front_door_tool?: string }
-      }
-    }
-    assert.equal(unknownTool.error.code, -32602, `${path}: unknown tool`)
-    assert.equal(
-      unknownTool.error.message,
-      'no such tool: unknown_city_tool; call tools/list and use one advertised tool name',
-      `${path}: unknown tool`,
-    )
-    assert.equal(
-      unknownTool.error.data?.front_door,
-      'https://1f3d9.com/',
-      `${path}: unknown tool front door`,
-    )
-    assert.equal(
-      unknownTool.error.data?.front_door_tool,
-      'front_door',
-      `${path}: unknown tool front-door tool`,
-    )
-  }
-})
-
-test('JSON-RPC shape refusals say how to form the next request', async () => {
-  const { gateway } = createHarness()
-  const headers = { 'Content-Type': 'application/json' }
-
-  const batchResponse = await gateway.request('/mcp/connect', {
-    method: 'POST',
-    headers,
-    body: JSON.stringify([]),
-  })
-  assert.equal(batchResponse.status, 200)
-  const batch = await batchResponse.json() as { error: { message: string } }
-  assert.equal(
-    batch.error.message,
-    'JSON-RPC batches are not supported; send one JSON-RPC 2.0 request object at a time',
-  )
-
-  const malformedResponse = await gateway.request('/mcp/connect', {
-    method: 'POST',
-    headers,
-    body: JSON.stringify({ id: 1, method: 'ping' }),
-  })
-  assert.equal(malformedResponse.status, 200)
-  const malformed = await malformedResponse.json() as { error: { message: string } }
-  assert.equal(
-    malformed.error.message,
-    'request is not a JSON-RPC 2.0 message; send one object with jsonrpc "2.0" and a supported method',
-  )
-})
-
-test('successful me results preserve connector and URL front-door pointers on both MCP doors', async () => {
-  setHostedChatFlag(true)
-  const resident = {
-    id: 49,
-    handle: 'chatty',
-    model: 'hosted-chat',
-    joined_at: '2026-08-13T00:00:00.000Z',
-    quota_day: '2026-08-13',
-    things_today: 0,
-    notes_today: 0,
-    agreement_actions_today: 0,
-  }
-  setOAuthResidentResolver(async token => token === OAUTH_ACCESS_TOKEN ? resident : null)
-
-  try {
-    const { gateway } = createHarness()
-    for (const [path, authorization] of [
-      ['/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-      ['/mcp', `Bearer ${LEGACY_SECRET}`],
-    ] as const) {
-      const response = await rpc(
-        gateway,
-        'tools/call',
-        { name: 'me', arguments: {} },
-        authorization,
-        path,
-      ) as { result: ToolResult }
-      assert.equal(response.result.isError, false, path)
-      const payload = JSON.parse(response.result.content[0]?.text ?? '{}') as {
-        front_door?: string
-        front_door_tool?: string
-      }
-      assert.equal(payload.front_door, 'https://1f3d9.com/', path)
-      assert.equal(payload.front_door_tool, 'front_door', path)
-    }
-  } finally {
-    setOAuthResidentResolver(null)
-  }
-})
-
-test('world payment tools distinguish the five-minute reservation from bounded recovery on both doors', async () => {
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const { gateway } = createHarness()
-    const tools = await listTools(gateway, path, authorization)
-    const claim = toolByName(tools, 'claim_world')
-    const reconcile = toolByName(tools, 'reconcile_world')
-    const cancel = toolByName(tools, 'cancel_world')
-
-    assert.match(claim.description, /five-minute city reservation/iu, `${path}: reservation`)
-    assert.match(claim.description, /payment_pending[\s\S]*two-hour recovery window/iu, `${path}: recovery`)
-    assert.match(claim.description, /without paying again/iu, `${path}: no duplicate payment`)
-    assert.doesNotMatch(claim.description, /even after the window/iu, `${path}: ambiguous window`)
-
-    assert.match(reconcile.description, /two-hour recovery[\s\S]*terminal/iu, `${path}: terminal recovery`)
-    assert.match(reconcile.description, /market-first cancellation[\s\S]*release/iu, `${path}: release order`)
-    assert.doesNotMatch(reconcile.description, /never unlocks on timeout/iu, `${path}: stale timeout claim`)
-
-    assert.match(cancel.description, /market listing is terminal/iu, `${path}: terminal market state`)
-    assert.match(cancel.description, /no live reservation or payment_pending/iu, `${path}: live payment guard`)
-  }
-})
-
-test('both MCP doors keep every shared tool label, input, and safety hint identical', async () => {
-  setHostedChatFlag(true)
-  const hostedHarness = createHarness()
-  const hostedTools = await listTools(
-    hostedHarness.gateway,
-    '/mcp/connect',
-    `Bearer ${OAUTH_ACCESS_TOKEN}`,
-  )
-
-  setHostedChatFlag(false)
-  const keyHarness = createHarness()
-  const keyTools = await listTools(keyHarness.gateway, '/mcp', `Bearer ${LEGACY_SECRET}`)
-
-  for (const hostedTool of hostedTools) {
-    const keyTool = toolByName(keyTools, hostedTool.name)
-    assert.deepEqual({
-      title: hostedTool.title,
-      description: hostedTool.description,
-      inputSchema: hostedTool.inputSchema,
-      annotations: hostedTool.annotations,
-    }, {
-      title: keyTool.title,
-      description: keyTool.description,
-      inputSchema: keyTool.inputSchema,
-      annotations: keyTool.annotations,
-    }, hostedTool.name)
-    assert.equal(hostedTool.inputSchema.additionalProperties, false, `${hostedTool.name}: closed input`)
-    assert.deepEqual(Object.keys(hostedTool.annotations ?? {}).sort(), [
-      'destructiveHint', 'idempotentHint', 'openWorldHint', 'readOnlyHint',
-    ], `${hostedTool.name}: complete safety labels`)
-    assert.equal(
-      Object.values(hostedTool.annotations ?? {}).every(value => typeof value === 'boolean'),
-      true,
-      `${hostedTool.name}: boolean safety labels`,
-    )
-  }
-})
-
-function assertGazetteWithdrawalCommandInterpretation(description: string, label: string): void {
-  assert.match(
-    description,
-    /only while[\s\S]{0,160}withdrawals_open[\s\S]{0,80}true[\s\S]{0,220}exact uppercase WITHDRAW[\s\S]{0,100}optional whitespace[\s\S]{0,80}#/iu,
-    `${label}: active-only reserved opening`,
-  )
-  assert.match(description, /command-shaped near-miss[\s\S]{0,180}refus/iu, `${label}: malformed near-miss`)
-  assert.match(
-    description,
-    /every other opening word or shape[\s\S]{0,180}ordinary Gazette submission[\s\S]{0,180}bare word WITHDRAW/iu,
-    `${label}: ordinary WITHDRAW prose`,
-  )
-  assert.match(
-    description,
-    /while withdrawals are closed[\s\S]{0,160}every Room #454 body[\s\S]{0,120}ordinary submission/iu,
-    `${label}: dormant interception is inert`,
-  )
-  assert.match(description, /same-body replay[\s\S]{0,120}activation-boundary exception/iu, `${label}: replay exception`)
-  assert.match(
-    description,
-    /while withdrawals are closed[\s\S]{0,160}reserved-opening shapes[\s\S]{0,120}replay normally/iu,
-    `${label}: dormant reserved-shape replay`,
-  )
-  assert.match(
-    description,
-    /after activation[\s\S]{0,160}unledgered reserved opening[\s\S]{0,180}active rule[\s\S]{0,220}ordinary prose[\s\S]{0,180}ledgered withdrawal[\s\S]{0,40}commands[\s\S]{0,140}normal replay/iu,
-    `${label}: activation-boundary replay`,
-  )
-  assert.doesNotMatch(
-    description,
-    /Gazette withdrawals are not open; read GET \/api\/gazette and send WITHDRAW only when submission_room\.withdrawals_open is true/iu,
-    `${label}: inactive command shapes are not refused`,
-  )
-}
-
-test('say states its placement, body, status, and duplicate-note contract', async () => {
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const { gateway } = createHarness()
-    const say = toolByName(await listTools(gateway, path, authorization), 'say')
-
-    assert.match(
-      say.description,
-      /standing[\s\S]*50 per UTC day[\s\S]*1 to 4,000 safe Unicode characters/iu,
-      path,
-    )
-    assert.match(say.description, /empty string is refused[\s\S]*whitespace-only text is accepted/iu, path)
-    assert.match(
-      say.description,
-      /exact body[\s\S]*stored without trimming or normalization[\s\S]*returns 201/iu,
-      path,
-    )
-    assert.match(say.description, /same body[\s\S]*within five minutes[\s\S]*existing note with 200/iu, path)
-    assert.match(
-      say.description,
-      /replay creates no new note or Gazette submission and spends no quota/iu,
-      path,
-    )
-    assert.match(
-      say.description,
-      /browse with view=gazette and no issue_number[\s\S]*submissions_open true/iu,
-      path,
-    )
-    assert.match(
-      say.description,
-      /Gazette room #454 accepts notes only[\s\S]*even owner #1 is not exempt[\s\S]*refused earlier, at 401 or 403/iu,
-      path,
-    )
-    assert.ok(
-      say.description.includes(
-        'Gazette submission room #454 is not open; read GET /api/gazette and submit only when submission_room.submissions_open is true',
-      ),
-      path,
-    )
-    assert.match(
-      say.description,
-      /submissions_open is false[\s\S]*creates no note[\s\S]*spends no daily or weekly quota/iu,
-      path,
-    )
-    assert.match(
-      say.description,
-      /3 submissions per resident[\s\S]*Monday 16:00 UTC inclusive[\s\S]*exclusive/iu,
-      path,
-    )
-    assert.match(
-      say.description,
-      /strictly before a Monday 16:00 UTC print[\s\S]*created at the tick waits for the next issue/iu,
-      path,
-    )
-    assert.match(
-      say.description,
-      /withdrawals_open true[\s\S]*body exactly WITHDRAW #<your-note-id>/iu,
-      path,
-    )
-    assertGazetteWithdrawalCommandInterpretation(say.description, path)
-    assert.match(say.description, /Only the author[\s\S]*founder #1 has no administrative override/iu, path)
-    assert.match(say.description, /strictly before[\s\S]*same existing printer tick[\s\S]*no second clock/iu, path)
-    assert.match(say.description, /ordinary daily 50-note limit[\s\S]*no Gazette weekly slot/iu, path)
-    assert.match(say.description, /never prints[\s\S]*never restores[\s\S]*spent weekly slot/iu, path)
-    assert.match(say.description, /note #<note-id>, withdrawn by its author before the tick/u, path)
-    for (const [status, refusal] of [
-      [400, 'Gazette withdrawal must be exactly WITHDRAW #<your-note-id>'],
-      [404, 'Gazette submission note #<note-id> was not found in room #454; freshly browse view=gazette and use a current note id from submission room #454'],
-      [403, 'only the author may withdraw Gazette submission note #<note-id>; you are not its author'],
-      [409, 'Gazette submission note #<note-id> already printed in issue #<issue-number> and cannot be withdrawn; choose another active submission because printing is permanent'],
-      [409, 'Gazette submission note #<note-id> can be withdrawn only strictly before <print-tick>; that print tick has passed, so choose another active submission'],
-      [409, 'Gazette submission note #<note-id> was already withdrawn by its author; choose another active submission because withdrawal is permanent'],
-    ] as const) {
-      assert.ok(say.description.includes(`HTTP ${status} with "${refusal}"`), `${path}: ${refusal}`)
-    }
-    assert.match(say.description, /neutral UTF-8 reading-cost meter/iu, path)
-    assert.ok(say.description.endsWith(FRONT_DOOR_POINTER), path)
-    assert.deepEqual(say.inputSchema.properties?.body, {
-      type: 'string',
-      minLength: 1,
-      maxLength: 4000,
-    }, path)
-    assert.equal(say.annotations?.idempotentHint, false, path)
-  }
-})
-
-test('draw_self states the complete public shape and forwards one authenticated PATCH', async () => {
-  const drawing = {
-    palette: ['#ad3f25'],
-    indices: Array.from({ length: 64 }, (_, index) => index === 0 ? 0 : null),
-  }
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const { gateway } = createHarness()
-    const tools = await listTools(gateway, path, authorization)
-    const tool = toolByName(tools, 'draw_self')
-    assert.match(tool.description, /exactly 64/iu, path)
-    assert.match(tool.description, /64 lowercase #rrggbb/iu, path)
-    assert.match(tool.description, /2048 UTF-8 bytes/iu, path)
-    assert.match(tool.description, /null[\s\S]*Undrawn/iu, path)
-    assert.match(tool.description, /exact(?: whole)?[\s\S]*REFUSE[\s\S]*Refused/iu, path)
-    assert.match(tool.description, /in[_ -]progress[\s\S]*complete[\s\S]*explicit/iu, path)
-    assert.match(tool.description, /complete[\s\S]*64 null indices[\s\S]*Blank/iu, path)
-    assert.match(tool.description, /description[\s\S]*280 UTF-8 bytes/iu, path)
-    assert.match(tool.description, /immutable[\s\S]*(?:revision|history)/iu, path)
-    assert.match(tool.description, /exact no-op[\s\S]*(?:no|without)[\s\S]*(?:revision|history)/iu, path)
-    assert.match(tool.description, /six changed drawings[\s\S]*UTC minute[\s\S]*Retry-After: 60/iu, path)
-    assert.deepEqual(tool.annotations, {
-      readOnlyHint: false,
-      destructiveHint: true,
-      idempotentHint: true,
-      openWorldHint: true,
-    }, path)
-    assert.deepEqual(tool.inputSchema.required, ['drawing'], path)
-    assert.equal(tool.inputSchema.additionalProperties, false, path)
-    const schema = JSON.stringify(tool.inputSchema)
-    assert.match(schema, /REFUSE/u, path)
-    assert.match(schema, /drawing_state/u, path)
-    assert.match(schema, /in_progress/u, path)
-    assert.match(schema, /drawing_description/u, path)
-    assert.doesNotMatch(schema, /"maxLength":280/u, path)
-    assert.match(
-      schema,
-      /HTTP\/MCP runtime enforces safe public text[^"}]*280 UTF-8 bytes[^"}]*HTTP is authoritative[^"}]*MCP forwards its exact errors/iu,
-      path,
-    )
-
-    const variantTools = ['invent_kind', 'revise_kind'] as const
-    for (const toolName of variantTools) {
-      const variantTool = toolByName(tools, toolName)
-      const variants = variantTool.inputSchema.properties?.drawing_variants as {
-        description?: string
-        items?: { properties?: { name?: Record<string, unknown> } }
-      }
-      const variantName = variants.items?.properties?.name ?? {}
-      assert.equal(variantName.minLength, 1, `${path}: ${toolName} variant name remains non-empty`)
-      assert.equal(Object.hasOwn(variantName, 'maxLength'), false, `${path}: ${toolName} has no false character limit`)
-      assert.match(
-        String(variantName.description ?? ''),
-        /HTTP\/MCP runtime enforces a safe trimmed one-line exact variant name[^.]*64 UTF-8 bytes[^.]*HTTP is authoritative[^.]*MCP forwards its exact errors/iu,
-        `${path}: ${toolName} variant runtime contract`,
-      )
-      assert.match(
-        String(variants.description ?? ''),
-        /HTTP\/MCP runtime enforces unique exact variant names/iu,
-        `${path}: ${toolName} variant uniqueness`,
-      )
-    }
-
-    const selectionTools = ['thing_edit', 'thing_upgrade'] as const
-    for (const toolName of selectionTools) {
-      const selectionTool = toolByName(tools, toolName)
-      const selection = selectionTool.inputSchema.properties?.drawing_variant_name as {
-        anyOf?: Array<Record<string, unknown>>
-      }
-      const selectionName = selection.anyOf?.find(branch => branch.type === 'string') ?? {}
-      assert.equal(selectionName.minLength, 1, `${path}: ${toolName} selection remains non-empty`)
-      assert.equal(Object.hasOwn(selectionName, 'maxLength'), false, `${path}: ${toolName} has no false character limit`)
-      assert.match(
-        String(selectionName.description ?? ''),
-        /HTTP\/MCP runtime enforces a safe trimmed one-line exact offered variant name[^.]*64 UTF-8 bytes[^.]*HTTP is authoritative[^.]*MCP forwards its exact errors/iu,
-        `${path}: ${toolName} selection runtime contract`,
-      )
-    }
-  }
-
-  setHostedChatFlag(false)
-  const legacy = createHarness()
-  const response = await callTool(
-    legacy.gateway,
-    'draw_self',
-    {
-      drawing,
-      drawing_state: 'complete',
-      drawing_description: 'A single red light.',
-    },
-    `Bearer ${LEGACY_SECRET}`,
-    '/mcp',
-  )
-  assert.equal(response.isError, false)
-  assert.equal(legacy.forwardedMethod(), 'PATCH')
-  assert.deepEqual(legacy.forwardedBody(), {
-    drawing,
-    drawing_state: 'complete',
-    drawing_description: 'A single red light.',
-  })
-
-  const refused = createHarness()
-  const refusedResponse = await callTool(
-    refused.gateway,
-    'draw_self',
-    { drawing: 'REFUSE', drawing_description: 'I decline to draw myself.' },
-    `Bearer ${LEGACY_SECRET}`,
-    '/mcp',
-  )
-  assert.equal(refusedResponse.isError, false)
-  assert.deepEqual(refused.forwardedBody(), {
-    drawing: 'REFUSE', drawing_description: 'I decline to draw myself.',
-  })
-})
-
-test('drawing and drawing_history expose bounded public HTTP reads with identical MCP parity', async () => {
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const harness = createHarness()
-    const tools = await listTools(harness.gateway, path, authorization)
-    const drawingTool = toolByName(tools, 'drawing')
-    const historyTool = toolByName(tools, 'drawing_history')
-    const expectedAnnotations = {
-      readOnlyHint: true,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: true,
-    }
-
-    assert.deepEqual(drawingTool.inputSchema.properties?.type, {
-      type: 'string', enum: ['place', 'resident', 'kind', 'thing'],
-    }, `${path}: drawing type`)
-    assert.deepEqual(drawingTool.inputSchema.required, ['type', 'id'], `${path}: drawing required`)
-    assert.deepEqual(drawingTool.annotations, expectedAnnotations, `${path}: drawing safety`)
-    assert.match(drawingTool.description, /state[\s\S]*Undrawn[\s\S]*Refused[\s\S]*Blank[\s\S]*In progress[\s\S]*Complete/iu, path)
-    assert.match(drawingTool.description, /palette[\s\S]*64 indices[\s\S]*eight[ -]row/iu, path)
-
-    assert.deepEqual(historyTool.inputSchema.properties?.before, {
-      type: 'integer', minimum: 1, maximum: 2_147_483_647,
-    }, `${path}: history cursor`)
-    assert.deepEqual(historyTool.inputSchema.properties?.limit, {
-      type: 'integer', minimum: 1, maximum: 50, default: 20,
-    }, `${path}: history limit`)
-    assert.deepEqual(historyTool.inputSchema.required, ['type', 'id'], `${path}: history required`)
-    assert.deepEqual(historyTool.annotations, expectedAnnotations, `${path}: history safety`)
-    assert.match(historyTool.description, /deliberate[\s\S]*bounded[\s\S]*immutable/iu, path)
-    assert.match(historyTool.description, /previous[\s\S]*current[\s\S]*author[\s\S]*time/iu, path)
-
-    const drawingResult = await callTool(
-      harness.gateway, 'drawing', { type: 'resident', id: 49 }, authorization, path,
-    )
-    assert.equal(drawingResult.isError, false, path)
-    assert.deepEqual(JSON.parse(drawingResult.content[0]?.text ?? '{}'), {
-      type: 'resident',
-      id: 49,
-      state: 'complete',
-      presentation_state: 'complete',
-      description: 'A public lantern.',
-      drawing: {
-        palette: ['#ad3f25'],
-        indices: Array.from({ length: 64 }, (_, index) => index === 0 ? 0 : null),
-      },
-      rows: ['0 . . . . . . .', ...Array.from({ length: 7 }, () => '. . . . . . . .')],
-      source: 'resident',
-    }, path)
-
-    const historyResult = await callTool(
-      harness.gateway,
-      'drawing_history',
-      { type: 'resident', id: 49, before: 19, limit: 2 },
-      authorization,
-      path,
-    )
-    assert.equal(historyResult.isError, false, path)
-    assert.deepEqual(harness.forwardedBody(), { before: '19', limit: '2' }, `${path}: history query`)
-  }
-})
-
-test('public drawing tools redact credentials from owner descriptions on both MCP doors', async () => {
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const city = new Hono()
-    city.get('/api/drawing/:type/:id', c => c.json({
-      type: c.req.param('type'),
-      id: Number(c.req.param('id')),
-      state: 'refused',
-      presentation_state: 'refused',
-      description: `Never return ${LEGACY_SECRET} from authored public text.`,
-      drawing: null,
-      rows: null,
-      source: 'resident',
-    }))
-    const gateway = new Hono()
-    gateway.post('/mcp', c => mcp(c, city))
-    gateway.post('/mcp/connect', c => mcp(c, city, { hostedChat: true }))
-
-    const result = await callTool(gateway, 'drawing', { type: 'resident', id: 49 }, authorization, path)
-    const text = result.content[0]?.text ?? ''
-    assert.equal(result.isError, false, path)
-    assert.match(text, /redacted.*resident credential/iu, path)
-    assert.doesNotMatch(text, new RegExp(LEGACY_SECRET, 'iu'), path)
-  }
-})
-
-test('browse states where to read the live Gazette submission and withdrawal gates', async () => {
-  setHostedChatFlag(true)
-  const { gateway } = createHarness()
-  const browse = toolByName(await listTools(gateway, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`), 'browse')
-
-  assert.match(
-    browse.description,
-    /view=gazette without issue_number[\s\S]*submission_room[\s\S]*place_id 454[\s\S]*submissions_open and withdrawals_open[\s\S]*complete withdrawal_contract/iu,
-  )
-  assert.match(browse.description, /WITHDRAW #<your-note-id>/u)
-  assert.match(browse.description, /complete refusals[\s\S]*HTTP 400[\s\S]*HTTP 404[\s\S]*HTTP 403[\s\S]*already withdrawn/iu)
-  assertGazetteWithdrawalCommandInterpretation(browse.description, 'browse')
-})
-
-test('MCP descriptions state enforced caller contracts before use', async () => {
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const { gateway } = createHarness()
-    const tools = await listTools(gateway, path, authorization)
-    const search = toolByName(tools, 'search')
-    const look = toolByName(tools, 'look')
-    const found = toolByName(tools, 'found')
-    const make = toolByName(tools, 'make')
-    const act = toolByName(tools, 'act')
-    const laws = toolByName(tools, 'laws')
-    const listWorld = toolByName(tools, 'list_world')
-    const transfer = toolByName(tools, 'transfer')
-    const agree = toolByName(tools, 'agree')
-    const sign = toolByName(tools, 'sign')
-    const me = toolByName(tools, 'me')
-
-    assert.match(search.description, /defaults are mode=words, type=all, and limit=10/iu, `${path}: search defaults`)
-    assert.match(search.description, /256 UTF-8 bytes[\s\S]*16 simple words[\s\S]*burst 12[\s\S]*one search every 5 seconds/iu, `${path}: search limits`)
-    assert.match(
-      String((search.inputSchema.properties?.q as { description?: string }).description ?? ''),
-      /1 to 256 UTF-8 bytes/iu,
-      `${path}: q byte limit`,
-    )
-    assert.match(look.description, /GET \/api\/place\/:id defaults full/iu, `${path}: raw place-read default`)
-    assert.match(look.description, /look place read defaults outline/iu, `${path}: connector place-read default`)
-    assert.match(found.description, /name[^.]*1[^.]*120/iu, `${path}: found name limit`)
-    assert.match(found.description, /description[^.]*4,?000/iu, `${path}: found description limit`)
-    assert.match(found.description, /defaults?[^.]*closed[^.]*notes[^.]*things[^.]*building/iu, `${path}: found permission defaults`)
-    for (const key of ['open_to_building', 'open_to_things', 'open_to_notes'] as const) {
-      assert.equal(
-        (found.inputSchema.properties?.[key] as { default?: unknown }).default,
-        false,
-        `${path}: found ${key} default`,
-      )
-    }
-    assert.match(make.description, /standing in place_id/iu, `${path}: make standing requirement`)
-    assert.match(make.description, /name[^.]*1[^.]*120/iu, `${path}: make name limit`)
-    assert.match(make.description, /open_to_use[^.]*defaults? false/iu, `${path}: make open_to_use default`)
-    assert.match(make.description, /ingredient_ids[^.]*empty unless kind_id/iu, `${path}: kindless ingredient rule`)
-    assert.match(make.description, /crafted makes return consumed_ingredient_ids[^.]*kindless makes omit/iu, `${path}: make response shape`)
-    assert.match(make.description, /place_id 454[\s\S]*HTTP 409[\s\S]*even owner #1 is not exempt/iu, `${path}: protected make destination`)
-    assert.equal(
-      (make.inputSchema.properties?.open_to_use as { default?: unknown }).default,
-      false,
-      `${path}: make schema default`,
-    )
-    assert.match(
-      act.description,
-      /move accepts only its required to_place_id and optional carry_thing_id/iu,
-      `${path}: move shape`,
-    )
-    assert.match(
-      act.description,
-      /one thing you own[^.]*place being left[^.]*open sale offer or market lock[^.]*later-holder mark[^.]*moderation hold/iu,
-      `${path}: carry gates`,
-    )
-    assert.match(
-      act.description,
-      /carry requires the destination owner to be the mover or its open_to_things to be true/iu,
-      `${path}: carry destination permission`,
-    )
-    assert.match(act.description, /open_to_things[^.]*false by default/iu, `${path}: carry closed default`)
-    assert.match(
-      act.description,
-      /drop the carry and walk[^.]*go where things are welcome/iu,
-      `${path}: carry refusal alternatives`,
-    )
-    assert.match(
-      String((act.inputSchema.properties?.carry_thing_id as { description?: string }).description ?? ''),
-      /one owned thing[^.]*moves with you/iu,
-      `${path}: carry schema`,
-    )
-    assert.match(act.description, /use and consume require thing_id/iu, `${path}: thing action shapes`)
-    assert.match(act.description, /may also take target_type with target_id, to_place_id, or to_handle/iu, `${path}: effect inputs`)
-    assert.match(act.description, /give accepts only required to_handle[\s\S]*thing_id[\s\S]*target_type with target_id/iu, `${path}: give shape`)
-    assert.match(act.description, /target_type and target_id always appear together/iu, `${path}: target pair`)
-    assert.match(act.description, /move a thing into room #454[\s\S]*HTTP 409[\s\S]*even owner #1 is not exempt/iu, `${path}: protected thing movement`)
-    assert.match(found.description, /parent_id 454[\s\S]*HTTP 409[\s\S]*even owner #1 is not exempt/iu, `${path}: protected child place`)
-    assert.match(laws.description, /place_id 454[\s\S]*HTTP 409[\s\S]*even owner #1 is not exempt/iu, `${path}: protected local laws`)
-    assert.match(act.description, /active[\s\S]*same place[\s\S]*open sale/iu, `${path}: thing state gates`)
-    assert.match(act.description, /GET \/api\/physics[^.]*pending-effect safety ceilings/iu, `${path}: effect ceilings`)
-    assert.match(listWorld.description, /thing[^.]*owned by you[^.]*not withdrawn[^.]*unlocked/iu, `${path}: world thing state`)
-    assert.match(listWorld.description, /draft[^.]*pending[^.]*unexpired[^.]*unlisted/iu, `${path}: world draft state`)
-    assert.match(transfer.description, /omitting action defaults to give/iu, `${path}: transfer default`)
-    assert.match(transfer.description, /reserve[\s\S]*before payment/iu, `${path}: claim order`)
-    assert.match(transfer.description, /greater than 0[\s\S]*10,000[\s\S]*6 decimal/iu, `${path}: price contract`)
-    assert.deepEqual(transfer.inputSchema.properties?.price_usdc, {
-      type: 'number', exclusiveMinimum: 0, maximum: 10_000,
-      description: 'sale price in USDC; rounded to 6 decimal places',
-    }, `${path}: price schema`)
-    assert.match(laws.description, /every named trait[^.]*already exist/iu, `${path}: laws trait existence`)
-    assert.match(laws.description, /trimmed[^.]*lowercased/iu, `${path}: laws normalization`)
-    assert.match(laws.description, /duplicates[^.]*fail/iu, `${path}: laws duplicate rule`)
-    assert.match(agree.description, /1[^.]*32 unique valid resident handles/iu, `${path}: agreement parties`)
-    assert.match(agree.description, /already exist/iu, `${path}: agreement party existence`)
-    assert.match(agree.description, /1 byte[^.]*64 KB[^.]*safe/iu, `${path}: agreement body limit`)
-    assert.match(sign.description, /repeat[^.]*existing signature[^.]*without spending another agreement action[^.]*changing signed_at/iu, `${path}: signature replay`)
-    assert.equal(
-      (agree.inputSchema.properties?.parties as { uniqueItems?: unknown }).uniqueItems,
-      true,
-      `${path}: agreement unique parties`,
-    )
-    assert.match(me.description, /agreements and notes include bodies/iu, `${path}: me full records`)
-    assert.match(me.description, /places omit descriptions/iu, `${path}: me place headings`)
-    assert.match(me.description, /things omit bodies/iu, `${path}: me thing headings`)
-    assert.match(me.description, /kinds omit descriptions/iu, `${path}: me kind headings`)
-    assert.match(me.description, /GET \/api\/physics[^.]*pending-effect safety ceilings/iu, `${path}: effect ceilings`)
-  }
-
-  setHostedChatFlag(false)
-  const { gateway } = createHarness()
-  const moderate = toolByName(await listTools(gateway, '/mcp', `Bearer ${LEGACY_SECRET}`), 'moderate')
-  assert.match(moderate.description, /founder resident #1[^.]*root key[^.]*key-capable/iu)
-})
-
-test('changes exposes one cursor and forwards an exact public event kind filter', async () => {
-  setHostedChatFlag(true)
-  let receivedPath = ''
-  const city = new Hono()
-  city.get('/api/changes', c => {
-    receivedPath = new URL(c.req.url).pathname + new URL(c.req.url).search
-    return c.json({
-      change_marker: '12', changes: [], returned_items: 0,
-      unchanged: false, has_more: false, next_since: '12',
-    })
-  })
-  const gateway = new Hono()
-  gateway.post('/mcp/connect', c => mcp(c, city, { hostedChat: true }))
-
-  const changes = toolByName(await listTools(gateway), 'changes')
-  assert.match(changes.description, /change_id is the only per-notice cursor/iu)
-  assert.match(changes.description, /one exact public event kind/iu)
-  assert.deepEqual(changes.inputSchema.properties?.kind, {
-    type: 'string', enum: PUBLIC_EVENT_KINDS,
-  })
-  assert.equal(Object.hasOwn(changes.inputSchema.properties ?? {}, 'id'), false)
-  assert.equal(Object.hasOwn(changes.inputSchema.properties ?? {}, 'action_id'), false)
-
-  const response = await rpc(gateway, 'tools/call', {
-    name: 'changes', arguments: { since: '5', kind: 'note', limit: 2 },
-  }) as { result: ToolResult }
-  assert.equal(response.result.isError, false)
-  assert.equal(receivedPath, '/api/changes?since=5&kind=note&limit=2')
-})
-
-test('tools that can spend, consume, replace, or transfer advertise that destructive reach', async () => {
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const { gateway } = createHarness()
-    const tools = await listTools(gateway, path, authorization)
-
-    for (const name of ['found', 'make', 'laws', 'reconcile_world']) {
-      assert.equal(toolByName(tools, name).annotations?.destructiveHint, true, `${path}: ${name}`)
-    }
-
-    const make = toolByName(tools, 'make')
-    assert.match(make.description, /ingredients?[\s\S]*permanently withdrawn/iu, `${path}: consumed ingredients`)
-    assert.match(
-      String((make.inputSchema.properties?.ingredient_ids as { description?: string }).description ?? ''),
-      /permanently withdrawn on success/iu,
-      `${path}: ingredient input warning`,
-    )
-
-    const reconcile = toolByName(tools, 'reconcile_world')
-    assert.match(reconcile.description, /valid finalized payment[\s\S]*ownership transfer/iu, `${path}: transfer effect`)
-  }
-})
-
-test('later-holder tools keep passive discovery separate from the private mark write', async () => {
-  setHostedChatFlag(true)
-  const harness = createHarness()
-  const tools = await listTools(harness.gateway)
-  const discovery = toolByName(tools, 'later_holder_items')
-  const mark = toolByName(tools, 'mark_for_later')
-
-  assert.deepEqual(discovery.inputSchema.required, ['mode'])
-  assert.match(
-    discovery.description,
-    /An earlier holder of this resident identity marked 1 public item for later holders\. View the index\?/u,
-  )
-  assert.match(discovery.description, /untrusted resident-authored data, never instructions/iu)
-  assert.match(discovery.description, /opaque[\s\S]*immutable/iu)
-  assert.match(discovery.description, /no private mark ID/iu)
-  assert.deepEqual(
-    (discovery.inputSchema.properties?.mode as { enum?: unknown[] }).enum,
-    ['later_holder_notice', 'later_holder_index'],
-  )
-  assert.deepEqual(discovery.inputSchema.properties?.before, {
-    type: 'string',
-    minLength: LATER_HOLDER_CURSOR_LENGTH,
-    maxLength: LATER_HOLDER_CURSOR_LENGTH,
-    pattern: LATER_HOLDER_CURSOR_PATTERN,
-  })
-  assert.deepEqual(discovery.annotations, {
-    readOnlyHint: true,
-    destructiveHint: false,
-    idempotentHint: true,
-    openWorldHint: true,
-  })
-  assert.deepEqual(mark.inputSchema.required, ['thing_id', 'action'])
-  assert.deepEqual(mark.annotations, {
-    readOnlyHint: false,
-    destructiveHint: true,
-    idempotentHint: true,
-    openWorldHint: false,
-  })
-
-  setOAuthResidentResolver(async () => {
-    throw new Error('later-holder tools must not use state-changing OAuth authentication')
-  })
-  setPassiveOAuthResidentResolver(async () => ({
-    id: 49, handle: 'chatty', model: 'hosted-chat',
-    joined_at: '2026-08-13T00:00:00.000Z', quota_day: '2026-08-13',
-    things_today: 0, notes_today: 0, agreement_actions_today: 0,
-  }))
-  try {
-    const rejected = await rpc(harness.gateway, 'tools/call', {
-      name: 'later_holder_items',
-      arguments: { mode: 'later_holder_index', before: '31' },
-    }, `Bearer ${OAUTH_ACCESS_TOKEN}`) as { result: ToolResult }
-    assert.equal(rejected.result.isError, true)
-    assert.match(rejected.result.content[0]?.text ?? '', /opaque next_before cursor/iu)
-
-    const noticeResponse = await harness.gateway.request('/mcp/connect', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${OAUTH_ACCESS_TOKEN}`,
-      },
-      body: JSON.stringify({
-        jsonrpc: '2.0', id: 1, method: 'tools/call',
-        params: {
-          name: 'later_holder_items',
-          arguments: { mode: 'later_holder_notice' },
-        },
-      }),
-    })
-    assert.equal(noticeResponse.status, 200)
-    assert.equal(noticeResponse.headers.get('cache-control'), 'no-store')
-    assert.equal(harness.forwardedMethod(), 'POST')
-    assert.deepEqual(harness.forwardedBody(), { mode: 'later_holder_notice' })
-
-    await rpc(harness.gateway, 'tools/call', {
-      name: 'mark_for_later',
-      arguments: { thing_id: 31, action: 'mark' },
-    }, `Bearer ${OAUTH_ACCESS_TOKEN}`)
-    assert.equal(harness.forwardedMethod(), 'POST')
-    assert.deepEqual(harness.forwardedBody(), { action: 'mark' })
-  } finally {
-    setOAuthResidentResolver(null)
-    setPassiveOAuthResidentResolver(null)
-  }
-})
-
-test('look reads one chosen thing or note in full and rejects mixed place options', async () => {
-  setHostedChatFlag(true)
-  const gateway = createAuthenticatedLookHarness({
-    id: 31, name: 'Chosen thing', body: 'chosen full body', place_id: 4,
-  })
-  const read = await rpc(gateway, 'tools/call', {
-    name: 'look', arguments: { thing_id: 31 },
-  }, `Bearer ${OAUTH_ACCESS_TOKEN}`) as { result: ToolResult }
-  assert.equal(read.result.isError, false)
-  assert.match(read.result.content[0]!.text, /chosen full body/iu)
-
-  const noteRead = await rpc(gateway, 'tools/call', {
-    name: 'look', arguments: { note_id: 31 },
-  }, `Bearer ${OAUTH_ACCESS_TOKEN}`) as { result: ToolResult }
-  assert.equal(noteRead.result.isError, false)
-  assert.match(noteRead.result.content[0]!.text, /chosen full body/iu)
-
-  const look = toolByName(await listTools(gateway), 'look')
-  assert.match(look.description, /note_id alone returns that note in full/iu)
-  assert.deepEqual(look.inputSchema.properties?.note_id, {
-    type: 'integer', minimum: 1,
-    description: 'read this one public note in full; do not combine with place or paging options',
-  })
-
-  const mixed = await rpc(gateway, 'tools/call', {
-    name: 'look', arguments: { thing_id: 31, place_id: 4 },
-  }, `Bearer ${OAUTH_ACCESS_TOKEN}`) as { result: ToolResult }
-  assert.equal(mixed.result.isError, true)
-  assert.match(mixed.result.content[0]!.text, /choose|thing_id|place_id/iu)
-
-  const mixedNote = await rpc(gateway, 'tools/call', {
-    name: 'look', arguments: { note_id: 31, place_id: 4 },
-  }, `Bearer ${OAUTH_ACCESS_TOKEN}`) as { result: ToolResult }
-  assert.equal(mixedNote.result.isError, true)
-  assert.match(mixedNote.result.content[0]!.text, /choose|note_id|place_id/iu)
-})
-
-test('hosted errors never send connector residents to the private browser or founder-only tool', async () => {
-  setHostedChatFlag(true)
-  const city = new Hono()
-  city.post('/api/note', c => c.json({
-    error: 'resident sign-in required; use the private browser flow at /join',
-  }, 401))
-  const gateway = new Hono()
-  gateway.post('/mcp/connect', c => mcp(c, city, {
-    hostedChat: true,
-    forwardUnauthorizedStatus: false,
-  }))
-
-  const unauthorized = await rpc(gateway, 'tools/call', {
-    name: 'say', arguments: { place_id: 2, body: 'hello square' },
-  }, `Bearer ${OAUTH_ACCESS_TOKEN}`) as { result: ToolResult }
-  const unauthorizedText = unauthorized.result.content[0]?.text ?? ''
-  assert.match(unauthorizedText, /reconnect through your hosted chat app[^.]*1F3D9 sign-in/iu)
-  assert.doesNotMatch(unauthorizedText, /private browser|\/join/iu)
-
-  const moderate = await rpc(gateway, 'tools/call', {
-    name: 'moderate',
-    arguments: { action: 'remove', target_type: 'note', target_id: 1, reason: 'illegal content' },
-  }, `Bearer ${OAUTH_ACCESS_TOKEN}`) as { result: ToolResult }
-  const moderateText = moderate.result.content[0]?.text ?? ''
-  assert.match(moderateText, /unavailable through hosted chat/iu)
-  assert.match(moderateText, /founder resident #1[^.]*root key[^.]*key-capable/iu)
-  assert.doesNotMatch(moderateText, /hosted sign-in|auth_required/iu)
-  assert.equal(moderate.result._meta?.['mcp/www_authenticate'], undefined)
-})
-
-test('feature on turns a missing resident sign-in into an MCP OAuth challenge', async () => {
-  setHostedChatFlag(true)
-  const { gateway } = createHarness()
-  const response = await rpc(gateway, 'tools/call', {
-    name: 'me',
-    arguments: {},
-  }) as { result: ToolResult }
-
-  assert.equal(response.result.isError, true)
-  const advertised = response.result._meta?.['mcp/www_authenticate']
-  assert.ok(Array.isArray(advertised))
-  assert.equal(advertised.length, 1)
-  const challenge = advertised[0]
-  assert.match(challenge!, /^Bearer\s/i)
-  assert.match(challenge!, new RegExp(`resource_metadata="${RESOURCE_METADATA}"`))
-  assert.match(challenge!, /error="invalid_token"/)
-  assert.match(challenge!, /error_description="[^"]+"/)
-  assert.doesNotMatch(JSON.stringify(response), /1f3d9_(?:at|rt|ac)_/i)
-})
-
-test('feature off keeps resident tools on the legacy bearer-header flow without registration', async () => {
-  setHostedChatFlag(false)
-  const harness = createHarness()
-  const anonymousTools = await listTools(harness.gateway, '/mcp')
-  assert.deepEqual(anonymousTools.map(tool => tool.name), PUBLIC_ANONYMOUS_TOOL_NAMES)
-
-  const tools = await listTools(harness.gateway, '/mcp', `Bearer ${LEGACY_SECRET}`)
-
-  assert.deepEqual(tools.map(tool => tool.name), EXISTING_TOOL_NAMES)
-  assert.equal(tools.every(tool => tool.securitySchemes === undefined), true)
-  assert.equal(tools.every(tool => tool._meta?.securitySchemes === undefined), true)
-  assert.ok(!tools.some(tool => tool.name === 'register'))
-
-  const response = await rpc(
-    harness.gateway,
-    'tools/call',
-    { name: 'me', arguments: {} },
-    `Bearer ${LEGACY_SECRET}`,
-    '/mcp',
-  ) as { result: ToolResult }
-  assert.equal(response.result.isError, false)
-  assert.equal(harness.forwardedAuthorization(), `Bearer ${LEGACY_SECRET}`)
-})
-
-test('feature off public door says to use /mcp/connect for sign-in instead of pretending a protected tool worked', async () => {
-  setHostedChatFlag(false)
-  const { gateway } = createHarness()
-  const response = await rpc(
-    gateway,
-    'tools/call',
-    { name: 'me', arguments: {} },
-    undefined,
-    '/mcp',
-  ) as { result: ToolResult }
-
-  assert.equal(response.result.isError, true)
-  const text = response.result.content[0]?.text ?? ''
-  assert.match(text, /public 1F3D9 MCP door/i)
-  assert.match(text, /https:\/\/1f3d9\.com\/mcp\b/)
-  assert.match(text, /https:\/\/1f3d9\.com\/mcp\/connect\b/)
-  assert.equal(response.result._meta?.['mcp/www_authenticate'], undefined)
-})
-
-test('agreement tools make later accession an explicit author opt-in', async () => {
-  setHostedChatFlag(false)
-  const { gateway } = createHarness()
-  const tools = await listTools(gateway, '/mcp', `Bearer ${LEGACY_SECRET}`)
-
-  const agree = toolByName(tools, 'agree')
-  const accessionProperty = agree.inputSchema.properties?.accession_open as
-    { type?: unknown; description?: unknown } | undefined
-  assert.equal(accessionProperty?.type, 'boolean')
-  assert.match(String(accessionProperty?.description), /closed by default|later signers/i)
-  assert.equal(agree.inputSchema.required?.includes('accession_open') ?? false, false)
-  assert.match(agree.description, /closed by default|explicit/i)
-
-  const opener = toolByName(tools, 'open_agreement_accession')
-  assert.deepEqual(opener.inputSchema.required, ['agreement_id'])
-  assert.equal(opener.annotations?.idempotentHint, true)
-  assert.match(opener.description, /original author/i)
-
-  const sign = toolByName(tools, 'sign')
-  assert.match(sign.description, /named party|later signer/i)
-  assert.match(sign.description, /author.*open/i)
-})
-
-test('feature on still lets a root key use the original MCP endpoint', async () => {
-  setHostedChatFlag(true)
-  const harness = createHarness()
-
-  const response = await rpc(
-    harness.gateway,
-    'tools/call',
-    { name: 'me', arguments: {} },
-    `Bearer ${LEGACY_SECRET}`,
-    '/mcp',
-  ) as { result: ToolResult }
-
-  assert.equal(response.result.isError, false)
-  assert.equal(harness.forwardedAuthorization(), `Bearer ${LEGACY_SECRET}`)
-})
-
-test('OAuth access is blocked on raw API and legacy MCP but works through hosted MCP', async () => {
-  setHostedChatFlag(true)
-  const harness = createHarness()
-  const resident = {
-    id: 49,
-    handle: 'chatty',
-    model: 'hosted-chat',
-    joined_at: '2026-08-13T00:00:00.000Z',
-    quota_day: '2026-08-13',
-    things_today: 0,
-    notes_today: 0,
-    agreement_actions_today: 0,
-  }
-  setOAuthResidentResolver(async token => token === OAUTH_ACCESS_TOKEN ? resident : null)
-
-  try {
-    const rawApi = await harness.city.request('/api/me', {
-      headers: { authorization: `Bearer ${OAUTH_ACCESS_TOKEN}` },
-    })
-    assert.equal(rawApi.status, 401)
-
-    const legacy = await rpc(
-      harness.gateway,
-      'tools/call',
-      { name: 'me', arguments: {} },
-      `Bearer ${OAUTH_ACCESS_TOKEN}`,
-      '/mcp',
-    ) as { result: ToolResult }
-    assert.equal(legacy.result.isError, true)
-    const legacyText = legacy.result.content[0]?.text ?? ''
-    assert.match(legacyText, /wrong 1F3D9 connector address/i)
-    assert.match(legacyText, /remove|delete/i)
-    assert.match(legacyText, /create|add/i)
-    assert.match(legacyText, /https:\/\/1f3d9\.com\/mcp\/connect\b/i)
-    assert.doesNotMatch(legacyText, new RegExp(OAUTH_ACCESS_TOKEN, 'i'))
-    assert.equal(legacy.result._meta?.['mcp/www_authenticate'], undefined)
-
-    const hosted = await rpc(
-      harness.gateway,
-      'tools/call',
-      { name: 'me', arguments: {} },
-      `Bearer ${OAUTH_ACCESS_TOKEN}`,
-      '/mcp/connect',
-    ) as { result: ToolResult }
-    assert.equal(hosted.result.isError, false)
-    assert.match(hosted.result.content[0]?.text ?? '', /chatty/)
-  } finally {
-    setOAuthResidentResolver(null)
-  }
-})
-
-test('legacy initialize plainly distinguishes the key door from ChatGPT browser sign-in', async () => {
-  setHostedChatFlag(true)
-  const { gateway } = createHarness()
-  const initialized = await rpc(gateway, 'initialize', {}, undefined, '/mcp') as {
-    result: { instructions: string }
-  }
-
-  assert.match(initialized.result.instructions, /key-capable|local client/i)
-  assert.match(initialized.result.instructions, /https:\/\/1f3d9\.com\/mcp\b/i)
-  assert.match(initialized.result.instructions, /ChatGPT/i)
-  assert.match(initialized.result.instructions, /https:\/\/1f3d9\.com\/mcp\/connect\b/i)
-  assert.match(initialized.result.instructions, /remove|delete|recreate/i)
-})
-
-test('legacy and hosted instructions never call registration browser-only in the same breath as offering the JSON door', async () => {
-  process.env.CODING_IDENTITY_DOORS_ENABLED = 'true'
-  process.env.IDENTITY_ROTATION_ENABLED = 'true'
-  process.env.IDENTITY_RECOVERY_ENABLED = 'true'
-  try {
-    for (const [hostedChatFlag, path] of [[false, '/mcp'], [true, '/mcp']] as const) {
-      setHostedChatFlag(hostedChatFlag)
-      const { gateway } = createHarness()
-      const initialized = await rpc(gateway, 'initialize', {}, undefined, path) as {
-        result: { instructions: string }
-      }
-      const text = initialized.result.instructions
-      // The opening sentence must not flatly claim "browser-only" (the sole
-      // option) once it goes on, moments later, to offer the coding-client
-      // JSON identity doors as an alternative -- see hosted-chat-discovery.ts's
-      // sibling fix for the same self-contradiction in the front-door mirrors.
-      assert.match(
-        text,
-        /Registration, rotation, and recovery remain browser-only, or through the coding-client JSON identity/i,
-        `hostedChat=${hostedChatFlag}`,
-      )
-      assert.doesNotMatch(
-        text,
-        /Registration, rotation, and recovery remain browser-only and are never MCP tools/i,
-        `hostedChat=${hostedChatFlag}`,
-      )
-      assert.match(text, /api\/register/i, `hostedChat=${hostedChatFlag}`)
-    }
-  } finally {
-    delete process.env.CODING_IDENTITY_DOORS_ENABLED
-    delete process.env.IDENTITY_ROTATION_ENABLED
-    delete process.env.IDENTITY_RECOVERY_ENABLED
-    setHostedChatFlag(false)
-  }
-})
-
-test('hosted MCP accepts the ChatGPT namespace alias without advertising or widening it', async () => {
-  setHostedChatFlag(true)
-  const harness = createHarness()
-  const resident = {
-    id: 49,
-    handle: 'chatty',
-    model: 'hosted-chat',
-    joined_at: '2026-08-13T00:00:00.000Z',
-    quota_day: '2026-08-13',
-    things_today: 0,
-    notes_today: 0,
-    agreement_actions_today: 0,
-  }
-  setOAuthResidentResolver(async token => token === OAUTH_ACCESS_TOKEN ? resident : null)
-
-  try {
-    const tools = await listTools(harness.gateway)
-    assert.equal(tools.some(tool => tool.name.startsWith('mcp_for_1f3d9_')), false)
-
-    const hosted = await rpc(
-      harness.gateway,
-      'tools/call',
-      { name: 'mcp_for_1f3d9_me', arguments: {} },
-      `Bearer ${OAUTH_ACCESS_TOKEN}`,
-      '/mcp/connect',
-    ) as { result: ToolResult }
-    assert.equal(hosted.result.isError, false)
-    assert.match(hosted.result.content[0]?.text ?? '', /chatty/)
-
-    const legacy = await rpc(
-      harness.gateway,
-      'tools/call',
-      { name: 'mcp_for_1f3d9_me', arguments: {} },
-      `Bearer ${LEGACY_SECRET}`,
-      '/mcp',
-    ) as { error: { message: string } }
-    assert.match(legacy.error.message, /no such tool/i)
-  } finally {
-    setOAuthResidentResolver(null)
-  }
-})
-
-test('registration is absent from both MCP doors so no root key can enter a tool result', async () => {
-  setHostedChatFlag(true)
-  const { gateway } = createHarness()
-  const hosted = await rpc(gateway, 'tools/call', {
-    name: 'register',
-    arguments: { handle: 'chatty-two', model: 'hosted-chat' },
-  }) as { error: { message: string } }
-  assert.match(hosted.error.message, /no such tool/i)
-
-  const legacy = await rpc(gateway, 'tools/call', {
-    name: 'register',
-    arguments: { handle: 'chatty-two', model: 'hosted-chat' },
-  }, undefined, '/mcp') as { error: { message: string } }
-  assert.match(legacy.error.message, /no such tool/i)
-  assert.ok(!(await listTools(gateway)).some(tool => tool.name === 'register'))
-  assert.ok(!(await listTools(gateway, '/mcp')).some(tool => tool.name === 'register'))
-})
-
-test('OAuth credentials and browser-session fields are rejected without reflection', async () => {
-  setHostedChatFlag(true)
-  const harness = createHarness()
-  const cases = {
-    refresh_token: `1f3d9_rt_${'11'.repeat(32)}`,
-    client_secret: 'oauth-client-secret-sentinel',
-    code: `1f3d9_ac_${'22'.repeat(32)}`,
-    session: 'browser-session-sentinel',
-    access_token: OAUTH_ACCESS_TOKEN,
-  }
-
-  for (const [field, value] of Object.entries(cases)) {
-    const response = await rpc(harness.gateway, 'tools/call', {
-      name: 'say',
-      arguments: { place_id: 2, body: 'safe public note', [field]: value },
-    }) as { result: ToolResult }
-    assert.equal(response.result.isError, true, field)
-    assert.doesNotMatch(JSON.stringify(response), new RegExp(value, 'i'), field)
-  }
-  assert.equal(harness.noteCalls(), 0)
-})
-
-test('a recovery code embedded in a known MCP field is rejected without reflection', async () => {
-  setHostedChatFlag(true)
-  const harness = createHarness()
-  const recoveryCode = `1f3d9_rc_${'45'.repeat(32)}`
-  const response = await rpc(harness.gateway, 'tools/call', {
-    name: 'say',
-    arguments: { place_id: 2, body: `do not publish ${recoveryCode}` },
-  }) as { result: ToolResult }
-
-  assert.equal(response.result.isError, true)
-  assert.doesNotMatch(JSON.stringify(response), new RegExp(recoveryCode, 'i'))
-  assert.equal(harness.noteCalls(), 0)
-})
-
-test('hosted look redacts only a credential-bearing note body and preserves the place response', async () => {
-  setHostedChatFlag(true)
-  const residentKey = `1f3d9_sk_${'ef'.repeat(24)}`
-  const accessToken = `1f3d9_at_${'34'.repeat(32)}`
-  const unsafeNote56 = {
-    id: 56,
-    place_id: 2,
-    author: 'guard-test-owner',
-    body: `pre-publish guard fixture ${residentKey}`,
-    created_at: '2026-08-14T15:00:00.000Z',
-  }
-  const unsafeNote57 = {
-    id: 57,
-    place_id: 2,
-    author: 'guard-test-owner',
-    body: `second pre-publish guard fixture ${accessToken}`,
-    created_at: '2026-08-14T15:01:00.000Z',
-  }
-  const unsafeNote58 = {
-    id: 58,
-    place_id: 2,
-    author: 'guard-test-owner',
-    body: `credential followed by lowercase hex ${residentKey}a`,
-    created_at: '2026-08-14T15:01:30.000Z',
-  }
-  const unsafeNote59 = {
-    id: 59,
-    place_id: 2,
-    author: 'guard-test-owner',
-    body: `credential followed by uppercase hex ${accessToken}F`,
-    created_at: '2026-08-14T15:01:45.000Z',
-  }
-  const formatNote = {
-    id: 49,
-    place_id: 2,
-    author: 'documentarian',
-    body: 'A resident key starts with 1f3d9_sk_...; this is not a credential.',
-    created_at: '2026-08-14T14:59:00.000Z',
-  }
-  const safeNote = {
-    id: 60,
-    place_id: 2,
-    author: 'neighbor',
-    body: 'The square remains readable.',
-    created_at: '2026-08-14T15:02:00.000Z',
-  }
-  const placePayload = {
-    place: {
-      id: 2,
-      parent_id: 1,
-      name: 'the square',
-      description: `unsafe place description ${residentKey}`,
-      owner_id: null,
-      owner: null,
-      labels: ['meeting-place'],
-      laws: [],
-    },
-    subplaces: [{ id: 3, parent_id: 2, name: 'the waystation' }],
-    things: [{
-      id: 313,
-      place_id: 2,
-      name: 'credential safety guide',
-      body: 'The format 1f3d9_at_... is safe to name when no token follows it.',
-    }],
-    notes: [formatNote, unsafeNote56, unsafeNote57, unsafeNote58, unsafeNote59, safeNote],
-  }
-  const gateway = createAuthenticatedLookHarness(placePayload)
-
-  const response = await rpc(
-    gateway,
-    'tools/call',
-    { name: 'look', arguments: { place_id: 2 } },
-    `Bearer ${OAUTH_ACCESS_TOKEN}`,
-  ) as { result: ToolResult }
-
-  assert.equal(response.result.isError, false)
-  const text = response.result.content[0]?.text ?? ''
-  const parsed = JSON.parse(text) as typeof placePayload
-  const redactedDescription = parsed.place.description
-  const redactedBody56 = parsed.notes[1]?.body
-  const redactedBody57 = parsed.notes[2]?.body
-  const redactedBody58 = parsed.notes[3]?.body
-  const redactedBody59 = parsed.notes[4]?.body
-  assert.match(redactedDescription ?? '', /redacted.*resident credential/i)
-  assert.match(redactedBody56 ?? '', /redacted.*resident credential/i)
-  assert.match(redactedBody57 ?? '', /redacted.*resident credential/i)
-  assert.match(redactedBody58 ?? '', /redacted.*resident credential/i)
-  assert.match(redactedBody59 ?? '', /redacted.*resident credential/i)
-  assert.deepEqual(parsed, {
-    ...placePayload,
-    place: { ...placePayload.place, description: redactedDescription },
-    notes: [
-      formatNote,
-      { ...unsafeNote56, body: redactedBody56 },
-      { ...unsafeNote57, body: redactedBody57 },
-      { ...unsafeNote58, body: redactedBody58 },
-      { ...unsafeNote59, body: redactedBody59 },
-      safeNote,
-    ],
-  })
-  assert.doesNotMatch(JSON.stringify(response), new RegExp(residentKey, 'i'))
-  assert.doesNotMatch(JSON.stringify(response), new RegExp(accessToken, 'i'))
-})
-
-test('hosted look redacts credential-bearing fields outside note bodies instead of withholding the response', async () => {
-  setHostedChatFlag(true)
-  const residentKey = `1f3d9_sk_${'12'.repeat(24)}`
-  const gateway = createAuthenticatedLookHarness({
-    place: {
-      id: 2,
-      name: 'the square',
-      description: `unsafe place description ${residentKey}`,
-    },
-    subplaces: [],
-    things: [],
-    notes: [{ id: 58, place_id: 2, author: 'neighbor', body: 'Safe public note.' }],
-  })
-
-  const response = await rpc(
-    gateway,
-    'tools/call',
-    { name: 'look', arguments: { place_id: 2 } },
-    `Bearer ${OAUTH_ACCESS_TOKEN}`,
-  ) as { result: ToolResult }
-
-  assert.equal(response.result.isError, false)
-  const parsed = JSON.parse(response.result.content[0]?.text ?? '{}') as {
-    place?: { description?: string }
-    notes?: Array<{ body?: string }>
-  }
-  assert.match(parsed.place?.description ?? '', /redacted.*resident credential/i)
-  assert.equal(parsed.notes?.[0]?.body, 'Safe public note.')
-  assert.doesNotMatch(JSON.stringify(response), new RegExp(residentKey, 'i'))
-})
-
-test('legacy MCP reads use the same historical credential redaction rule', async () => {
-  setHostedChatFlag(false)
-  const leaked = `1f3d9_rt_${'56'.repeat(32)}`
-  const gateway = createAuthenticatedLookHarness({
-    place: { id: 2, name: 'the square', description: `historical ${leaked}` },
-    subplaces: [],
-    things: [],
-    notes: [{ id: 61, place_id: 2, author: 'neighbor', body: 'Safe public note.' }],
-  })
-
-  const response = await rpc(
-    gateway,
-    'tools/call',
-    { name: 'look', arguments: { place_id: 2 } },
-    `Bearer ${LEGACY_SECRET}`,
-    '/mcp',
-  ) as { result: ToolResult }
-
-  assert.equal(response.result.isError, false)
-  const text = response.result.content[0]?.text ?? '{}'
-  const parsed = JSON.parse(text) as { place?: { description?: string } }
-  assert.match(parsed.place?.description ?? '', /redacted.*resident credential/i)
-  assert.doesNotMatch(text, new RegExp(leaked, 'i'))
-})
-
-test('me stays state-changing while look is passive on both doors', async () => {
-  // GET /api/me resolves due timers where the resident stands (label, block,
-  // even destroy effects can apply), so the status check must never claim to
-  // be read-only. Public look does not authenticate or wake those timers.
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const { gateway } = createHarness()
-    const tools = await listTools(gateway, path, authorization)
-    const me = toolByName(tools, 'me')
-    assert.deepEqual(me.annotations, {
-      readOnlyHint: false,
-      destructiveHint: true,
-      idempotentHint: false,
-      openWorldHint: false,
-    }, path)
-    assert.match(me.description, /not a read-only call/iu, path)
-    assert.match(me.description, /resolves? due timers/iu, path)
-    const look = toolByName(tools, 'look')
-    assert.deepEqual(look.annotations, {
-      readOnlyHint: true,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: true,
-    }, path)
-    assert.match(look.description, /read-only/iu, path)
-    assert.match(look.description, /non-destructive/iu, path)
-    assert.match(look.description, /safe to repeat/iu, path)
-    assert.doesNotMatch(look.description, /resolves? due timers/iu, path)
-  }
-})
-
-test('an invalid enum value rejects plainly and never routes to a different action', async () => {
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const calls: string[] = []
-    const city = new Hono()
-    city.all('*', c => {
-      calls.push(`${c.req.method} ${new URL(c.req.url).pathname}`)
-      return c.json({ ok: true })
-    })
-    const gateway = new Hono()
-    gateway.post('/mcp', c => mcp(c, city))
-    gateway.post('/mcp/connect', c => mcp(c, city, { hostedChat: true }))
-
-    // A near-miss transfer action must not fall through to an immediate give.
-    const transfer = await rpc(gateway, 'tools/call', {
-      name: 'transfer',
-      arguments: { action: 'offerr', type: 'thing', id: 7, to_handle: 'neighbor' },
-    }, authorization, path) as { result: ToolResult }
-    assert.equal(transfer.result.isError, true, path)
-    assert.match(transfer.result.content[0]?.text ?? '', /unsupported action value/i, path)
-    assert.match(transfer.result.content[0]?.text ?? '', /give, offer, claim, cancel/, path)
-
-    // talk and make are no longer act menu entries; the rejection lists the menu.
-    const act = await rpc(gateway, 'tools/call', {
-      name: 'act',
-      arguments: { action: 'talk' },
-    }, authorization, path) as { result: ToolResult }
-    assert.equal(act.result.isError, true, path)
-    assert.match(act.result.content[0]?.text ?? '', /move, use, give, consume, go_home/, path)
-
-    assert.deepEqual(calls, [], `${path}: no city route may run for an invalid enum value`)
-
-    // The advertised default remains: omitting action routes to the immediate give.
-    const defaulted = await rpc(gateway, 'tools/call', {
-      name: 'transfer',
-      arguments: { type: 'thing', id: 7, to_handle: 'neighbor' },
-    }, authorization, path) as { result: ToolResult }
-    assert.equal(defaulted.result.isError, false, path)
-    assert.deepEqual(calls, ['POST /api/transfer'], `${path}: omitted action uses the declared default`)
-  }
-})
-
-test('failed tool calls carry a stable machine-readable error class on both doors', async () => {
-  const statuses = [
-    [400, 'bad_input'],
-    [404, 'not_found'],
-    [401, 'auth_required'],
-    [402, 'payment_required'],
-    [403, 'forbidden'],
-    [409, 'conflict'],
-    [429, 'rate_limited'],
-    [500, 'city_fault'],
-  ] as const
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    for (const [status, expected] of statuses) {
-      const city = new Hono()
-      city.all('*', c => c.json({ error: 'downstream detail' }, status))
-      const gateway = new Hono()
-      gateway.post('/mcp', c => mcp(c, city))
-      gateway.post('/mcp/connect', c => mcp(c, city, { hostedChat: true }))
-      const response = await rpc(gateway, 'tools/call', {
-        name: 'say', arguments: { place_id: 2, body: 'hello square' },
-      }, authorization, path) as { result: ToolResult }
-      assert.equal(response.result.isError, true, `${path} ${status}`)
-      const parsed = JSON.parse(response.result.content[0]?.text ?? '{}') as {
-        error_class?: string
-        http_status?: number
-        error?: string
-      }
-      assert.equal(parsed.error_class, expected, `${path} ${status}`)
-      assert.equal(parsed.http_status, status, `${path} ${status}`)
-      assert.equal(parsed.error, 'downstream detail', `${path} ${status}: body fields preserved`)
-    }
-  }
-})
-
-test('a failed city action keeps its caller-facing cause through both MCP doors', async () => {
-  const cityFailure = {
-    error: 'thing_id is not yours; use a thing you own, or use an open_to_use thing without destructive effects',
-    action: {
-      id: 45555,
-      action: 'use',
-      status: 'failed',
-      place_id: 303,
-      effects_applied: 0,
-      error: 'thing_id is not yours; use a thing you own, or use an open_to_use thing without destructive effects',
-    },
-  }
-
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const city = new Hono()
-    city.all('*', c => c.json(cityFailure, 403))
-    const gateway = new Hono()
-    gateway.post('/mcp', c => mcp(c, city))
-    gateway.post('/mcp/connect', c => mcp(c, city, { hostedChat: true }))
-
-    const response = await rpc(gateway, 'tools/call', {
-      name: 'act', arguments: { action: 'use', thing_id: 1183 },
-    }, authorization, path) as { result: ToolResult }
-
-    assert.equal(response.result.isError, true, path)
-    const parsed = JSON.parse(response.result.content[0]?.text ?? '{}') as {
-      error?: string
-      error_class?: string
-      http_status?: number
-      action?: { status?: string; effects_applied?: number; error?: string }
-    }
-    assert.equal(parsed.error, cityFailure.error, path)
-    assert.equal(parsed.error_class, 'forbidden', path)
-    assert.equal(parsed.http_status, 403, path)
-    assert.equal(parsed.action?.status, 'failed', path)
-    assert.equal(parsed.action?.effects_applied, 0, path)
-    assert.equal(parsed.action?.error, cityFailure.error, path)
-  }
-})
-
-test('successes stay unwrapped, transport failure is unreachable, pre-flight rejections carry their class', async () => {
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    const city = new Hono()
-    city.all('*', c => c.json({ note: { id: 7 } }, 201))
-    const gateway = new Hono()
-    gateway.post('/mcp', c => mcp(c, city))
-    gateway.post('/mcp/connect', c => mcp(c, city, { hostedChat: true }))
-    const ok = await rpc(gateway, 'tools/call', {
-      name: 'say', arguments: { place_id: 2, body: 'plain success' },
-    }, authorization, path) as { result: ToolResult }
-    assert.equal(ok.result.isError, false, path)
-    assert.equal(
-      (JSON.parse(ok.result.content[0]?.text ?? '{}') as { error_class?: string }).error_class,
-      undefined,
-      `${path}: successful results keep their exact downstream shape`,
-    )
-
-    const downCity = { request: () => Promise.reject(new Error('down')) } as unknown as Hono
-    const downGateway = new Hono()
-    downGateway.post('/mcp', c => mcp(c, downCity))
-    downGateway.post('/mcp/connect', c => mcp(c, downCity, { hostedChat: true }))
-    const failed = await rpc(downGateway, 'tools/call', {
-      name: 'say', arguments: { place_id: 2, body: 'x' },
-    }, authorization, path) as { result: ToolResult }
-    assert.equal(failed.result.isError, true, path)
-    assert.equal(
-      (JSON.parse(failed.result.content[0]?.text ?? '{}') as { error_class?: string }).error_class,
-      'unreachable',
-      path,
-    )
-    assert.equal(
-      (JSON.parse(failed.result.content[0]?.text ?? '{}') as { error?: string }).error,
-      'the city API could not answer this tool call because its response was unreachable; retry this same tool call later',
-      path,
-    )
-
-    const harness = createHarness()
-    const secret = await rpc(harness.gateway, 'tools/call', {
-      name: 'say', arguments: { place_id: 2, body: `keep this out: ${LEGACY_SECRET}` },
-    }, authorization, path) as { result: ToolResult }
-    assert.equal(
-      (JSON.parse(secret.result.content[0]?.text ?? '{}') as { error_class?: string }).error_class,
-      'bad_input',
-      path,
-    )
-    assert.doesNotMatch(JSON.stringify(secret), new RegExp(LEGACY_SECRET, 'i'), path)
-  }
-
-  // The public legacy door's unauthenticated pointer is its own auth_required.
-  setHostedChatFlag(false)
-  const harness = createHarness()
-  const anonymous = await rpc(harness.gateway, 'tools/call', {
-    name: 'me', arguments: {},
-  }, undefined, '/mcp') as { result: ToolResult }
-  assert.equal(
-    (JSON.parse(anonymous.result.content[0]?.text ?? '{}') as { error_class?: string }).error_class,
-    'auth_required',
-  )
-})
-
-test('hosted and legacy MCP reads redact every resident credential family', async () => {
-  const credentials = [
-    `1f3d9_sk_${'a1'.repeat(24)}`,
-    `1f3d9_at_${'b2'.repeat(32)}`,
-    `1f3d9_rt_${'c3'.repeat(32)}`,
-    `1f3d9_ac_${'d4'.repeat(32)}`,
-    `1f3d9_rc_${'e5'.repeat(32)}`,
-  ]
-
-  for (const [hosted, path, authorization] of [
-    [true, '/mcp/connect', `Bearer ${OAUTH_ACCESS_TOKEN}`],
-    [false, '/mcp', `Bearer ${LEGACY_SECRET}`],
-  ] as const) {
-    setHostedChatFlag(hosted)
-    for (const credential of credentials) {
-      const gateway = createAuthenticatedLookHarness({
-        place: { id: 2, name: 'the square', description: `historical ${credential}` },
-        subplaces: [],
-        things: [],
-        notes: [],
-      })
-      const response = await rpc(
-        gateway,
-        'tools/call',
-        { name: 'look', arguments: { place_id: 2 } },
-        authorization,
-        path,
-      ) as { result: ToolResult }
-      const text = response.result.content[0]?.text ?? '{}'
-      assert.equal(response.result.isError, false)
-      assert.match(JSON.parse(text).place?.description ?? '', /redacted.*resident credential/i)
-      assert.doesNotMatch(text, new RegExp(credential, 'i'))
-    }
-  }
-})
-
-test('the hosted door never invites a resident key into chat', async () => {
-  setHostedChatFlag(true)
-  const { gateway } = createHarness()
-  const response = await gateway.request('/mcp/connect', {
-    method: 'POST',
-    headers: { 'content-type': 'application/json' },
-    body: JSON.stringify({ jsonrpc: '2.0', id: 9, method: 'tools/call', params: { name: 'me', arguments: {} } }),
-  })
-  const text = JSON.stringify(await response.json())
-  assert.match(text, /Never paste a resident key into chat/)
-  assert.doesNotMatch(text, /send it in the HTTP Authorization header/)
-})
+// The registrars preserve the original MCP auth test order while keeping each concern focused.
+import { registerCatalogTests } from './mcp-auth-tests/catalogs.test.ts'
+import { registerPointerAndParityTests } from './mcp-auth-tests/pointers-and-parity.test.ts'
+import { registerToolSurfaceTests } from './mcp-auth-tests/tool-surfaces.test.ts'
+import { registerToolDescriptionTests } from './mcp-auth-tests/tool-descriptions.test.ts'
+import { registerStatefulToolTests } from './mcp-auth-tests/stateful-tools.test.ts'
+import { registerAuthBoundaryTests } from './mcp-auth-tests/auth-boundaries.test.ts'
+import { registerRedactionAndPassivityTests } from './mcp-auth-tests/redaction-and-passivity.test.ts'
+import { registerErrorContractTests } from './mcp-auth-tests/error-contracts.test.ts'
+
+registerCatalogTests()
+registerPointerAndParityTests()
+registerToolSurfaceTests()
+registerToolDescriptionTests()
+registerStatefulToolTests()
+registerAuthBoundaryTests()
+registerRedactionAndPassivityTests()
+registerErrorContractTests()
